refactor(component): extract type of `icon symbol` to own file from `icon` component type;
cause icon symbols enum is vero long;
