+++ src/components/Icon/symbol.enum.ts
@@ -1,0 +1,2470 @@
+export enum IconSymbol {
+	Icon123 = '123',
+	Icon360 = '360',
+	'10k' = '10k',
+	'10mp' = '10mp',
+	'11mp' = '11mp',
+	'12mp' = '12mp',
+	'13mp' = '13mp',
+	'14mp' = '14mp',
+	'15mp' = '15mp',
+	'16mp' = '16mp',
+	'17mp' = '17mp',
+	'18UpRating' = '18-up-rating',
+	'18mp' = '18mp',
+	'19mp' = '19mp',
+	'1k' = '1k',
+	'1kPlus' = '1k-plus',
+	'1xMobileData' = '1x-mobiledata',
+	'20mp' = '20mp',
+	'21mp' = '21mp',
+	'22mp' = '22mp',
+	'23mp' = '23mp',
+	'24mp' = '24mp',
+	'2k' = '2k',
+	'2kPlus' = '2k-plus',
+	'2mp' = '2mp',
+	'30fps' = '30fps',
+	'30fpsSelect' = '30fps-select',
+	'3dRotation' = '3d-rotation',
+	'3gMobileData' = '3g-mobiledata',
+	'3k' = '3k',
+	'3kPlus' = '3k-plus',
+	'3mp' = '3mp',
+	'3p' = '3p',
+	'4gMobileData' = '4g-mobiledata',
+	'4gPlusMobileData' = '4g-plus-mobiledata',
+	'4k' = '4k',
+	'4kPlus' = '4k-plus',
+	'4mp' = '4mp',
+	'5g' = '5g',
+	'5k' = '5k',
+	'5kPlus' = '5k-plus',
+	'5mp' = '5mp',
+	'60fps' = '60fps',
+	'60fpsSelect' = '60fps-select',
+	'6FtApart' = '6-ft-apart',
+	'6k' = '6k',
+	'6kPlus' = '6k-plus',
+	'6mp' = '6mp',
+	'7k' = '7k',
+	'7kPlus' = '7k-plus',
+	'7mp' = '7mp',
+	'8k' = '8k',
+	'8kPlus' = '8k-plus',
+	'8mp' = '8mp',
+	'9k' = '9k',
+	'9kPlus' = '9k-plus',
+	'9mp' = '9mp',
+	Abc = 'abc',
+	AcUnit = 'ac-unit',
+	Accessibility = 'accessibility',
+	AccessibilityNew = 'accessibility-new',
+	Accessible = 'accessible',
+	AccessibleForward = 'accessible-forward',
+	AccountBalance = 'account-balance',
+	AccountBalanceWallet = 'account-balance-wallet',
+	AccountBox = 'account-box',
+	AccountCircle = 'account-circle',
+	AccountTree = 'account-tree',
+	ActivityZone = 'activity-zone',
+	Acute = 'acute',
+	AdUnits = 'ad-units',
+	Adb = 'adb',
+	Add = 'add',
+	AddAPhoto = 'add-a-photo',
+	AddAlert = 'add-alert',
+	AddBox = 'add-box',
+	AddBusiness = 'add-business',
+	AddCall = 'add-call',
+	AddCard = 'add-card',
+	AddChart = 'add-chart',
+	AddCircle = 'add-circle',
+	AddComment = 'add-comment',
+	AddHome = 'add-home',
+	AddHomeWork = 'add-home-work',
+	AddLink = 'add-link',
+	AddLocation = 'add-location',
+	AddLocationAlt = 'add-location-alt',
+	AddModerator = 'add-moderator',
+	AddNotes = 'add-notes',
+	AddPhotoAlternate = 'add-photo-alternate',
+	AddReaction = 'add-reaction',
+	AddRoad = 'add-road',
+	AddShoppingCart = 'add-shopping-cart',
+	AddTask = 'add-task',
+	AddToDrive = 'add-to-drive',
+	AddToHomeScreen = 'add-to-home-screen',
+	AddToPhotos = 'add-to-photos',
+	AddToQueue = 'add-to-queue',
+	AdfScanner = 'adf-scanner',
+	Adjust = 'adjust',
+	AdminMeds = 'admin-meds',
+	AdminPanelSettings = 'admin-panel-settings',
+	AdsClick = 'ads-click',
+	Agender = 'agender',
+	Agriculture = 'agriculture',
+	Air = 'air',
+	AirFreshener = 'air-freshener',
+	AirPurifierGen = 'air-purifier-gen',
+	AirlineSeatFlat = 'airline-seat-flat',
+	AirlineSeatFlatAngled = 'airline-seat-flat-angled',
+	AirlineSeatIndividualSuite = 'airline-seat-individual-suite',
+	AirlineSeatLegroomExtra = 'airline-seat-legroom-extra',
+	AirlineSeatLegroomNormal = 'airline-seat-legroom-normal',
+	AirlineSeatLegroomReduced = 'airline-seat-legroom-reduced',
+	AirlineSeatReclineExtra = 'airline-seat-recline-extra',
+	AirlineSeatReclineNormal = 'airline-seat-recline-normal',
+	AirlineStops = 'airline-stops',
+	Airlines = 'airlines',
+	AirplaneTicket = 'airplane-ticket',
+	AirplaneModeActive = 'airplanemode-active',
+	AirplaneModeInactive = 'airplanemode-inactive',
+	Airplay = 'airplay',
+	AirportShuttle = 'airport-shuttle',
+	AirWare = 'air-ware',
+	Airwave = 'airwave',
+	Alarm = 'alarm',
+	AlarmAdd = 'alarm-add',
+	AlarmOff = 'alarm-off',
+	AlarmOn = 'alarm-on',
+	Album = 'album',
+	AlignHorizontalCenter = 'align-horizontal-center',
+	AlignHorizontalLeft = 'align-horizontal-left',
+	AlignHorizontalRight = 'align-horizontal-right',
+	AlignVerticalBottom = 'align-vertical-bottom',
+	AlignVerticalCenter = 'align-vertical-center',
+	AlignVerticalTop = 'align-vertical-top',
+	AllInbox = 'all-inbox',
+	AllInclusive = 'all-inclusive',
+	AllOut = 'all-out',
+	Allergies = 'allergies',
+	AltRoute = 'alt-route',
+	AlternateEmail = 'alternate-email',
+	Altitude = 'altitude',
+	Amend = 'amend',
+	AmpStories = 'amp-stories',
+	Analytics = 'analytics',
+	Anchor = 'anchor',
+	Android = 'android',
+	Animation = 'animation',
+	Aod = 'aod',
+	AodTablet = 'aod-tablet',
+	Apartment = 'apartment',
+	Api = 'api',
+	ApkDocument = 'apk-document',
+	ApkInstall = 'apk-install',
+	AppBlocking = 'app-blocking',
+	AppRegistration = 'app-registration',
+	AppShortcut = 'app-shortcut',
+	Approval = 'approval',
+	ApprovalDelegation = 'approval-delegation',
+	Apps = 'apps',
+	AppsOutage = 'apps-outage',
+	Architecture = 'architecture',
+	Archive = 'archive',
+	AreaChart = 'area-chart',
+	ArmingCountdown = 'arming-countdown',
+	ArrowBack = 'arrow-back',
+	ArrowBackIos = 'arrow-back-ios',
+	ArrowBackIosNew = 'arrow-back-ios-new',
+	ArrowCircleDown = 'arrow-circle-down',
+	ArrowCircleLeft = 'arrow-circle-left',
+	ArrowCircleRight = 'arrow-circle-right',
+	ArrowCircleUp = 'arrow-circle-up',
+	ArrowDownward = 'arrow-downward',
+	ArrowDropDown = 'arrow-drop-down',
+	ArrowDropDownCircle = 'arrow-drop-down-circle',
+	ArrowDropUp = 'arrow-drop-up',
+	ArrowForward = 'arrow-forward',
+	ArrowForwardIos = 'arrow-forward-ios',
+	ArrowInsert = 'arrow-insert',
+	ArrowLeft = 'arrow-left',
+	ArrowOutward = 'arrow-outward',
+	ArrowRight = 'arrow-right',
+	ArrowRightAlt = 'arrow-right-alt',
+	ArrowSelectorTool = 'arrow-selector-tool',
+	ArrowUpward = 'arrow-upward',
+	ArrowsMoreDown = 'arrows-more-down',
+	ArrowsMoreUp = 'arrows-more-up',
+	ArtTrack = 'art-track',
+	Article = 'article',
+	AspectRatio = 'aspect-ratio',
+	Assignment = 'assignment',
+	AssignmentAdd = 'assignment-add',
+	AssignmentInd = 'assignment-ind',
+	AssignmentLate = 'assignment-late',
+	AssignmentReturn = 'assignment-return',
+	AssignmentReturned = 'assignment-returned',
+	AssignmentTurnedIn = 'assignment-turned-in',
+	AssistWalker = 'assist-walker',
+	Assistant = 'assistant',
+	AssistantDirection = 'assistant-direction',
+	AssistantNavigation = 'assistant-navigation',
+	AssuredWorkload = 'assured-workload',
+	AstrophotographyAuto = 'astrophotography-auto',
+	AstrophotographyOff = 'astrophotography-off',
+	ATM = 'atm',
+	Atm = ATM,
+	Atr = 'atr',
+	AttachEmail = 'attach-email',
+	AttachFile = 'attach-file',
+	AttachFileAdd = 'attach-file-add',
+	AttachMoney = 'attach-money',
+	Attachment = 'attachment',
+	Attractions = 'attractions',
+	Attribution = 'attribution',
+	AudioFile = 'audio-file',
+	AutoActivityZone = 'auto-activity-zone',
+	AutoAwesome = 'auto-awesome',
+	AutoAwesomeMosaic = 'auto-awesome-mosaic',
+	AutoAwesomeMotion = 'auto-awesome-motion',
+	AutoDelete = 'auto-delete',
+	AutoDetectVoice = 'auto-detect-voice',
+	AutoFix = 'auto-fix',
+	AutoFixHigh = 'auto-fix-high',
+	AutoFixNormal = 'auto-fix-normal',
+	AutoFixOff = 'auto-fix-off',
+	AutoGraph = 'auto-graph',
+	AutoMode = 'auto-mode',
+	AutoReadPause = 'auto-read-pause',
+	AutoReadPlay = 'auto-read-play',
+	AutoSchedule = 'auto-schedule',
+	AutoStories = 'auto-stories',
+	AutoFpsSelect = 'auto-fps-select',
+	AutoPay = 'auto-pay',
+	AutoRenew = 'auto-renew',
+	AvTimer = 'av-timer',
+	AvgTime = 'avg-time',
+	BabyChangingStation = 'baby-changing-station',
+	BackHand = 'back-hand',
+	BackgroundReplace = 'background-replace',
+	Backpack = 'backpack',
+	Backspace = 'backspace',
+	Backup = 'backup',
+	BackupTable = 'backup-table',
+	Badge = 'badge',
+	BadgeCriticalBattery = 'badge-critical-battery',
+	BakeryDining = 'bakery-dining',
+	Balance = 'balance',
+	Balcony = 'balcony',
+	Ballot = 'ballot',
+	BarChart = 'bar-chart',
+	Barcode = 'barcode',
+	BarcodeReader = 'barcode-reader',
+	BarcodeScanner = 'barcode-scanner',
+	Barefoot = 'barefoot',
+	BatchPrediction = 'batch-prediction',
+	Bathroom = 'bathroom',
+	Bathtub = 'bathtub',
+	Battery_0Bar = 'battery-0-bar',
+	Battery_1Bar = 'battery-1-bar',
+	Battery_20 = 'battery-20',
+	Battery_2Bar = 'battery-2-bar',
+	Battery_30 = 'battery-30',
+	Battery_3Bar = 'battery-3-bar',
+	Battery_4Bar = 'battery-4-bar',
+	Battery_50 = 'battery-50',
+	Battery_5Bar = 'battery-5-bar',
+	Battery_60 = 'battery-60',
+	Battery_6Bar = 'battery-6-bar',
+	Battery_80 = 'battery-80',
+	Battery_90 = 'battery-90',
+	BatteryAlert = 'battery-alert',
+	BatteryCharging_20 = 'battery-charging-20',
+	BatteryCharging_30 = 'battery-charging-30',
+	BatteryCharging_50 = 'battery-charging-50',
+	BatteryCharging_60 = 'battery-charging-60',
+	BatteryCharging_80 = 'battery-charging-80',
+	BatteryCharging_90 = 'battery-charging-90',
+	BatteryChargingFull = 'battery-charging-full',
+	BatteryFull = 'battery-full',
+	BatteryFullAlt = 'battery-full-alt',
+	BatteryHoriz_000 = 'battery-horiz-000',
+	BatteryHoriz_050 = 'battery-horiz-050',
+	BatteryHoriz_075 = 'battery-horiz-075',
+	BatteryLow = 'battery-low',
+	BatteryProfile = 'battery-profile',
+	BatterySaver = 'battery-saver',
+	BatteryUnknown = 'battery-unknown',
+	BatteryVert_005 = 'battery-vert-005',
+	BatteryVert_020 = 'battery-vert-020',
+	BatteryVert_050 = 'battery-vert-050',
+	BatteryVeryLow = 'battery-very-low',
+	BeachAccess = 'beach-access',
+	Bed = 'bed',
+	BedroomBaby = 'bedroom-baby',
+	BedroomChild = 'bedroom-child',
+	BedroomParent = 'bedroom-parent',
+	Bedtime = 'bedtime',
+	BedtimeOff = 'bedtime-off',
+	Beenhere = 'beenhere',
+	Bento = 'bento',
+	BikeScooter = 'bike-scooter',
+	Biotech = 'biotech',
+	Blanket = 'blanket',
+	Blender = 'blender',
+	Blind = 'blind',
+	Blinds = 'blinds',
+	BlindsClosed = 'blinds-closed',
+	Block = 'block',
+	Bloodtype = 'bloodtype',
+	Bluetooth = 'bluetooth',
+	BluetoothConnected = 'bluetooth-connected',
+	BluetoothDisabled = 'bluetooth-disabled',
+	BluetoothDrive = 'bluetooth-drive',
+	BluetoothSearching = 'bluetooth-searching',
+	BlurCircular = 'blur-circular',
+	BlurLinear = 'blur-linear',
+	BlurMedium = 'blur-medium',
+	BlurOff = 'blur-off',
+	BlurOn = 'blur-on',
+	BlurShort = 'blur-short',
+	BodyFat = 'body-fat',
+	Bolt = 'bolt',
+	Book = 'book',
+	BookOnline = 'book-online',
+	Bookmark = 'bookmark',
+	BookmarkAdd = 'bookmark-add',
+	BookmarkAdded = 'bookmark-added',
+	BookmarkRemove = 'bookmark-remove',
+	Bookmarks = 'bookmarks',
+	BorderAll = 'border-all',
+	BorderBottom = 'border-bottom',
+	BorderClear = 'border-clear',
+	BorderColor = 'border-color',
+	BorderHorizontal = 'border-horizontal',
+	BorderInner = 'border-inner',
+	BorderLeft = 'border-left',
+	BorderOuter = 'border-outer',
+	BorderRight = 'border-right',
+	BorderStyle = 'border-style',
+	BorderTop = 'border-top',
+	BorderVertical = 'border-vertical',
+	Boy = 'boy',
+	BrandingWatermark = 'branding-watermark',
+	BreakfastDining = 'breakfast-dining',
+	BreakingNewsAlt1 = 'breaking-news-alt-1',
+	Breastfeeding = 'breastfeeding',
+	Brightness1 = 'brightness-1',
+	Brightness2 = 'brightness-2',
+	Brightness3 = 'brightness-3',
+	Brightness4 = 'brightness-4',
+	Brightness5 = 'brightness-5',
+	Brightness6 = 'brightness-6',
+	Brightness7 = 'brightness-7',
+	BrightnessAuto = 'brightness-auto',
+	BrightnessHigh = 'brightness-high',
+	BrightnessLow = 'brightness-low',
+	BrightnessMedium = 'brightness-medium',
+	BringYourOwnIp = 'bring-your-own-ip',
+	BroadcastOnHome = 'broadcast-on-home',
+	BroadcastOnPersonal = 'broadcast-on-personal',
+	BrokenImage = 'broken-image',
+	BrowseActivity = 'browse-activity',
+	BrowseGallery = 'browse-gallery',
+	BrowserUpdated = 'browser-updated',
+	BrunchDining = 'brunch-dining',
+	Brush = 'brush',
+	BubbleChart = 'bubble-chart',
+	BugReport = 'bug-report',
+	Build = 'build',
+	BuildCircle = 'build-circle',
+	Bungalow = 'bungalow',
+	BurstMode = 'burst-mode',
+	BusAlert = 'bus-alert',
+	BusinessCenter = 'business-center',
+	BusinessChip = 'business-chip',
+	Cabin = 'cabin',
+	Cable = 'cable',
+	Cached = 'cached',
+	Cake = 'cake',
+	CakeAdd = 'cake-add',
+	Calculate = 'calculate',
+	CalendarAddOn = 'calendar-add-on',
+	CalendarAppsScript = 'calendar-apps-script',
+	CalendarMonth = 'calendar-month',
+	CalendarToday = 'calendar-today',
+	CalendarViewDay = 'calendar-view-day',
+	CalendarViewMonth = 'calendar-view-month',
+	CalendarViewWeek = 'calendar-view-week',
+	Call = 'call',
+	CallEnd = 'call-end',
+	CallMade = 'call-made',
+	CallMerge = 'call-merge',
+	CallMissed = 'call-missed',
+	CallMissedOutgoing = 'call-missed-outgoing',
+	CallReceived = 'call-received',
+	CallSplit = 'call-split',
+	CallToAction = 'call-to-action',
+	Camera = 'camera',
+	CameraEnhance = 'camera-enhance',
+	CameraFront = 'camera-front',
+	CameraIndoor = 'camera-indoor',
+	CameraOutdoor = 'camera-outdoor',
+	CameraRear = 'camera-rear',
+	CameraRoll = 'camera-roll',
+	CameraSwitch = 'camera-switch',
+	Campaign = 'campaign',
+	Camping = 'camping',
+	Cancel = 'cancel',
+	CancelPresentation = 'cancel-presentation',
+	CancelScheduleSend = 'cancel-schedule-send',
+	CandlestickChart = 'candlestick-chart',
+	CarCrash = 'car-crash',
+	CarRental = 'car-rental',
+	CarRepair = 'car-repair',
+	CardMembership = 'card-membership',
+	CardTravel = 'card-travel',
+	Carpenter = 'carpenter',
+	Cases = 'cases',
+	Casino = 'casino',
+	Cast = 'cast',
+	CastConnected = 'cast-connected',
+	CastForEducation = 'cast-for-education',
+	Castle = 'castle',
+	Category = 'category',
+	Celebration = 'celebration',
+	CellMerge = 'cell-merge',
+	CellTower = 'cell-tower',
+	CellWifi = 'cell-wifi',
+	CenterFocusStrong = 'center-focus-strong',
+	CenterFocusWeak = 'center-focus-weak',
+	Chair = 'chair',
+	ChairAlt = 'chair-alt',
+	Chalet = 'chalet',
+	ChangeCircle = 'change-circle',
+	ChangeHistory = 'change-history',
+	Charger = 'charger',
+	ChargingStation = 'charging-station',
+	ChartData = 'chart-data',
+	Chat = 'chat',
+	ChatAddOn = 'chat-add-on',
+	ChatAppsScript = 'chat-apps-script',
+	ChatBubble = 'chat-bubble',
+	Check = 'check',
+	CheckBox = 'check-box',
+	CheckBoxOutlineBlank = 'check-box-outline-blank',
+	CheckCircle = 'check-circle',
+	CheckIndeterminateSmall = 'check-indeterminate-small',
+	CheckSmall = 'check-small',
+	Checklist = 'checklist',
+	ChecklistRtl = 'checklist-rtl',
+	Checkroom = 'checkroom',
+	ChevronLeft = 'chevron-left',
+	ChevronRight = 'chevron-right',
+	ChildCare = 'child-care',
+	ChildFriendly = 'child-friendly',
+	ChipExtraction = 'chip-extraction',
+	ChromeReaderMode = 'chrome-reader-mode',
+	Chromecast2 = 'chromecast-2',
+	ChromecastDevice = 'chromecast-device',
+	Church = 'church',
+	CinematicBlur = 'cinematic-blur',
+	Circle = 'circle',
+	CircleNotifications = 'circle-notifications',
+	Clarify = 'clarify',
+	CleanHands = 'clean-hands',
+	CleaningBucket = 'cleaning-bucket',
+	CleaningServices = 'cleaning-services',
+	ClearAll = 'clear-all',
+	ClearDay = 'clear-day',
+	ClearNight = 'clear-night',
+	ClimateMiniSplit = 'climate-mini-split',
+	ClinicalNotes = 'clinical-notes',
+	Close = 'close',
+	CloseFullscreen = 'close-fullscreen',
+	ClosedCaption = 'closed-caption',
+	ClosedCaptionDisabled = 'closed-caption-disabled',
+	Cloud = 'cloud',
+	CloudCircle = 'cloud-circle',
+	CloudDone = 'cloud-done',
+	CloudDownload = 'cloud-download',
+	CloudOff = 'cloud-off',
+	CloudQueue = 'cloud-queue',
+	CloudSync = 'cloud-sync',
+	CloudUpload = 'cloud-upload',
+	Cloudy = 'cloudy',
+	CloudyFilled = 'cloudy-filled',
+	CloudySnowing = 'cloudy-snowing',
+	Co2 = 'co2',
+	CoPresent = 'co-present',
+	Code = 'code',
+	CodeBlocks = 'code-blocks',
+	CodeOff = 'code-off',
+	Coffee = 'coffee',
+	CoffeeMaker = 'coffee-maker',
+	Cognition = 'cognition',
+	CollectionsBookmark = 'collections-bookmark',
+	Colorize = 'colorize',
+	Comment = 'comment',
+	CommentBank = 'comment-bank',
+	CommentsDisabled = 'comments-disabled',
+	Commit = 'commit',
+	Communication = 'communication',
+	Commute = 'commute',
+	Compare = 'compare',
+	CompareArrows = 'compare-arrows',
+	CompassCalibration = 'compass-calibration',
+	ComponentExchange = 'component-exchange',
+	Compost = 'compost',
+	Compress = 'compress',
+	Computer = 'computer',
+	Conditions = 'conditions',
+	ConfirmationNumber = 'confirmation-number',
+	ConnectWithoutContact = 'connect-without-contact',
+	ConnectedTv = 'connected-tv',
+	ConnectingAirports = 'connecting-airports',
+	Construction = 'construction',
+	ContactEmergency = 'contact-emergency',
+	ContactMail = 'contact-mail',
+	ContactPage = 'contact-page',
+	ContactPhone = 'contact-phone',
+	ContactSupport = 'contact-support',
+	Contactless = 'contactless',
+	ContactlessOff = 'contactless-off',
+	Contacts = 'contacts',
+	ContentCopy = 'content-copy',
+	ContentCut = 'content-cut',
+	ContentPaste = 'content-paste',
+	ContentPasteGo = 'content-paste-go',
+	ContentPasteOff = 'content-paste-off',
+	ContentPasteSearch = 'content-paste-search',
+	Contrast = 'contrast',
+	ControlCamera = 'control-camera',
+	ControlPointDuplicate = 'control-point-duplicate',
+	ControllerGen = 'controller-gen',
+	ConversionPath = 'conversion-path',
+	ConveyorBelt = 'conveyor-belt',
+	Cookie = 'cookie',
+	Cooking = 'cooking',
+	CoolToDry = 'cool-to-dry',
+	CopyAll = 'copy-all',
+	Copyright = 'copyright',
+	Coronavirus = 'coronavirus',
+	CorporateFare = 'corporate-fare',
+	Cottage = 'cottage',
+	Countertops = 'countertops',
+	CreateNewFolder = 'create-new-folder',
+	CreditCard = 'credit-card',
+	CreditCardOff = 'credit-card-off',
+	CreditScore = 'credit-score',
+	Crib = 'crib',
+	CrisisAlert = 'crisis-alert',
+	Crop = 'crop',
+	Crop_16_9 = 'crop-16-9',
+	Crop_3_2 = 'crop-3-2',
+	Crop_5_4 = 'crop-5-4',
+	Crop_7_5 = 'crop-7-5',
+	CropFree = 'crop-free',
+	CropLandscape = 'crop-landscape',
+	CropPortrait = 'crop-portrait',
+	CropRotate = 'crop-rotate',
+	CropSquare = 'crop-square',
+	CrueltyFree = 'cruelty-free',
+	Css = 'css',
+	CurrencyBitcoin = 'currency-bitcoin',
+	CurrencyExchange = 'currency-exchange',
+	CurrencyFranc = 'currency-franc',
+	CurrencyLira = 'currency-lira',
+	CurrencyPound = 'currency-pound',
+	CurrencyRuble = 'currency-ruble',
+	CurrencyRupee = 'currency-rupee',
+	CurrencyYen = 'currency-yen',
+	CurrencyYuan = 'currency-yuan',
+	Curtains = 'curtains',
+	CurtainsClosed = 'curtains-closed',
+	Cut = 'cut',
+	Cycle = 'cycle',
+	Cyclone = 'cyclone',
+	Dangerous = 'dangerous',
+	DarkMode = 'dark-mode',
+	Dashboard = 'dashboard',
+	DashboardCustomize = 'dashboard-customize',
+	DataArray = 'data-array',
+	DataExploration = 'data-exploration',
+	DataObject = 'data-object',
+	DataSaverOn = 'data-saver-on',
+	DataThresholding = 'data-thresholding',
+	DataUsage = 'data-usage',
+	Database = 'database',
+	Dataset = 'dataset',
+	DatasetLinked = 'dataset-linked',
+	DateRange = 'date-range',
+	Deblur = 'deblur',
+	DecimalDecrease = 'decimal-decrease',
+	DecimalIncrease = 'decimal-increase',
+	Deck = 'deck',
+	Dehaze = 'dehaze',
+	Delete = 'delete',
+	DeleteForever = 'delete-forever',
+	DeleteSweep = 'delete-sweep',
+	Demography = 'demography',
+	DensityLarge = 'density-large',
+	DensityMedium = 'density-medium',
+	DensitySmall = 'density-small',
+	Dentistry = 'dentistry',
+	DepartureBoard = 'departure-board',
+	Dermatology = 'dermatology',
+	Description = 'description',
+	Deselect = 'deselect',
+	DesignServices = 'design-services',
+	Desk = 'desk',
+	DesktopAccessDisabled = 'desktop-access-disabled',
+	DesktopMac = 'desktop-mac',
+	DesktopWindows = 'desktop-windows',
+	Details = 'details',
+	DetectionAndZone = 'detection-and-zone',
+	Detector = 'detector',
+	DetectorAlarm = 'detector-alarm',
+	DetectorBattery = 'detector-battery',
+	DetectorCo = 'detector-co',
+	DetectorOffline = 'detector-offline',
+	DetectorSmoke = 'detector-smoke',
+	DetectorStatus = 'detector-status',
+	DeveloperBoard = 'developer-board',
+	DeveloperBoardOff = 'developer-board-off',
+	DeveloperMode = 'developer-mode',
+	DeviceHub = 'device-hub',
+	DeviceReset = 'device-reset',
+	DeviceThermostat = 'device-thermostat',
+	DeviceUnknown = 'device-unknown',
+	Devices = 'devices',
+	DevicesFold = 'devices-fold',
+	DevicesOther = 'devices-other',
+	DewPoint = 'dew-point',
+	Diagnosis = 'diagnosis',
+	DialerSip = 'dialer-sip',
+	Dialpad = 'dialpad',
+	Diamond = 'diamond',
+	Difference = 'difference',
+	DigitalOutOfHome = 'digital-out-of-home',
+	Dining = 'dining',
+	DinnerDining = 'dinner-dining',
+	Directions = 'directions',
+	DirectionsAlt = 'directions-alt',
+	DirectionsAltOff = 'directions-alt-off',
+	DirectionsBike = 'directions-bike',
+	DirectionsBoat = 'directions-boat',
+	DirectionsBus = 'directions-bus',
+	DirectionsCar = 'directions-car',
+	DirectionsOff = 'directions-off',
+	DirectionsRailway = 'directions-railway',
+	DirectionsRun = 'directions-run',
+	DirectionsSubway = 'directions-subway',
+	DirectionsWalk = 'directions-walk',
+	DirtyLens = 'dirty-lens',
+	DisabledByDefault = 'disabled-by-default',
+	DisabledVisible = 'disabled-visible',
+	DiscFull = 'disc-full',
+	DiscoverTune = 'discover-tune',
+	DishwasherGen = 'dishwasher-gen',
+	DisplaySettings = 'display-settings',
+	Diversity_1 = 'diversity-1',
+	Diversity_2 = 'diversity-2',
+	Diversity_3 = 'diversity-3',
+	Diversity_4 = 'diversity-4',
+	Dns = 'dns',
+	DoNotDisturbOff = 'do-not-disturb-off',
+	DoNotDisturbOn = 'do-not-disturb-on',
+	DoNotDisturbOnTotalSilence = 'do-not-disturb-on-total-silence',
+	DoNotStep = 'do-not-step',
+	DoNotTouch = 'do-not-touch',
+	Dock = 'dock',
+	DocsAddOn = 'docs-add-on',
+	DocsAppsScript = 'docs-apps-script',
+	DocumentScanner = 'document-scanner',
+	Domain = 'domain',
+	DomainAdd = 'domain-add',
+	DomainDisabled = 'domain-disabled',
+	DomainVerification = 'domain-verification',
+	Done = 'done',
+	DoneAll = 'done-all',
+	DoneOutline = 'done-outline',
+	DonutLarge = 'donut-large',
+	DonutSmall = 'donut-small',
+	DoorBack = 'door-back',
+	DoorFront = 'door-front',
+	DoorOpen = 'door-open',
+	DoorSensor = 'door-sensor',
+	DoorSliding = 'door-sliding',
+	Doorbell = 'doorbell',
+	Doorbell_3p = 'doorbell-3p',
+	DoorbellChime = 'doorbell-chime',
+	DoubleArrow = 'double-arrow',
+	DownhillSkiing = 'downhill-skiing',
+	Download = 'download',
+	DownloadDone = 'download-done',
+	DownloadForOffline = 'download-for-offline',
+	Downloading = 'downloading',
+	Draft = 'draft',
+	DraftOrders = 'draft-orders',
+	Drafts = 'drafts',
+	DragHandle = 'drag-handle',
+	DragIndicator = 'drag-indicator',
+	Draw = 'draw',
+	Dresser = 'dresser',
+	DriveFileMove = 'drive-file-move',
+	DriveFileMoveOutline = 'drive-file-move-outline',
+	DriveFileMoveRtl = 'drive-file-move-rtl',
+	DriveFileRenameOutline = 'drive-file-rename-outline',
+	DriveFolderUpload = 'drive-folder-upload',
+	Dry = 'dry',
+	DryCleaning = 'dry-cleaning',
+	Duo = 'duo',
+	Dvr = 'dvr',
+	DynamicFeed = 'dynamic-feed',
+	DynamicForm = 'dynamic-form',
+	E911Avatar = 'e911-avatar',
+	E911Emergency = 'e911-emergency',
+	EMobileData = 'e-mobiledata',
+	Earbuds = 'earbuds',
+	EarbudsBattery = 'earbuds-battery',
+	EarlyOn = 'early-on',
+	East = 'east',
+	Eco = 'eco',
+	EdgeSensorHigh = 'edge-sensor-high',
+	EdgeSensorLow = 'edge-sensor-low',
+	Edit = 'edit',
+	EditAttributes = 'edit-attributes',
+	EditCalendar = 'edit-calendar',
+	EditDocument = 'edit-document',
+	EditLocation = 'edit-location',
+	EditLocationAlt = 'edit-location-alt',
+	EditNote = 'edit-note',
+	EditNotifications = 'edit-notifications',
+	EditOff = 'edit-off',
+	EditRoad = 'edit-road',
+	EditSquare = 'edit-square',
+	Egg = 'egg',
+	EggAlt = 'egg-alt',
+	Eject = 'eject',
+	Elderly = 'elderly',
+	ElderlyWoman = 'elderly-woman',
+	ElectricBike = 'electric-bike',
+	ElectricBolt = 'electric-bolt',
+	ElectricCar = 'electric-car',
+	ElectricMeter = 'electric-meter',
+	ElectricMoped = 'electric-moped',
+	ElectricRickshaw = 'electric-rickshaw',
+	ElectricScooter = 'electric-scooter',
+	ElectricalServices = 'electrical-services',
+	Elevator = 'elevator',
+	Emergency = 'emergency',
+	EmergencyHeat = 'emergency-heat',
+	EmergencyHome = 'emergency-home',
+	EmergencyRecording = 'emergency-recording',
+	EmergencyShare = 'emergency-share',
+	EmojiFlags = 'emoji-flags',
+	EmojiFoodBeverage = 'emoji-food-beverage',
+	EmojiNature = 'emoji-nature',
+	EmojiObjects = 'emoji-objects',
+	EmojiPeople = 'emoji-people',
+	EmojiSymbols = 'emoji-symbols',
+	EmojiTransportation = 'emoji-transportation',
+	EmptyDashboard = 'empty-dashboard',
+	Enable = 'enable',
+	EnergyProgramSaving = 'energy-program-saving',
+	EnergyProgramTimeUsed = 'energy-program-time-used',
+	EnergySavingsLeaf = 'energy-savings-leaf',
+	Engineering = 'engineering',
+	EnhancedEncryption = 'enhanced-encryption',
+	Ent = 'ent',
+	Equalizer = 'equalizer',
+	Error = 'error',
+	ErrorCircleRounded = 'error-circle-rounded',
+	ErrorMed = 'error-med',
+	Escalator = 'escalator',
+	EscalatorWarning = 'escalator-warning',
+	Euro = 'euro',
+	EuroSymbol = 'euro-symbol',
+	EvCharger = 'ev-charger',
+	EvShadow = 'ev-shadow',
+	EvStation = 'ev-station',
+	Event = 'event',
+	EventAvailable = 'event-available',
+	EventBusy = 'event-busy',
+	EventNote = 'event-note',
+	EventRepeat = 'event-repeat',
+	EventSeat = 'event-seat',
+	EventUpcoming = 'event-upcoming',
+	Exclamation = 'exclamation',
+	ExitToApp = 'exit-to-app',
+	Expand = 'expand',
+	ExpandCircleDown = 'expand-circle-down',
+	ExpandContent = 'expand-content',
+	ExpandLess = 'expand-less',
+	ExpandMore = 'expand-more',
+	Explicit = 'explicit',
+	Explore = 'explore',
+	ExploreOff = 'explore-off',
+	ExportNotes = 'export-notes',
+	Exposure = 'exposure',
+	ExposureNeg_1 = 'exposure-neg-1',
+	ExposureNeg_2 = 'exposure-neg-2',
+	ExposurePlus_1 = 'exposure-plus-1',
+	ExposurePlus_2 = 'exposure-plus-2',
+	ExposureZero = 'exposure-zero',
+	Extension = 'extension',
+	ExtensionOff = 'extension-off',
+	Face = 'face',
+	Face_2 = 'face-2',
+	Face_3 = 'face-3',
+	Face_4 = 'face-4',
+	Face_5 = 'face-5',
+	Face_6 = 'face-6',
+	FaceRetouchingNatural = 'face-retouching-natural',
+	FaceRetouchingOff = 'face-retouching-off',
+	FactCheck = 'fact-check',
+	Factory = 'factory',
+	FamiliarFaceAndZone = 'familiar-face-and-zone',
+	FamilyHistory = 'family-history',
+	FamilyRestroom = 'family-restroom',
+	FastForward = 'fast-forward',
+	FastRewind = 'fast-rewind',
+	FastFood = 'fast-food',
+	Faucet = 'faucet',
+	Favorite = 'favorite',
+	Fax = 'fax',
+	FeaturedPlayList = 'featured-play-list',
+	FeaturedVideo = 'featured-video',
+	Feed = 'feed',
+	Female = 'female',
+	Femur = 'femur',
+	FemurAlt = 'femur-alt',
+	Fence = 'fence',
+	Festival = 'festival',
+	FiberDvr = 'fiber-dvr',
+	FiberManualRecord = 'fiber-manual-record',
+	FiberNew = 'fiber-new',
+	FiberPin = 'fiber-pin',
+	FiberSmartRecord = 'fiber-smart-record',
+	FileCopy = 'file-copy',
+	FileDownloadDone = 'file-download-done',
+	FileDownloadOff = 'file-download-off',
+	FileOpen = 'file-open',
+	FilePresent = 'file-present',
+	FileUploadOff = 'file-upload-off',
+	Filter = 'filter',
+	Filter1 = 'filter-1',
+	Filter2 = 'filter-2',
+	Filter3 = 'filter-3',
+	Filter4 = 'filter-4',
+	Filter5 = 'filter-5',
+	Filter6 = 'filter-6',
+	Filter7 = 'filter-7',
+	Filter8 = 'filter-8',
+	Filter9 = 'filter-9',
+	Filter9Plus = 'filter-9-plus',
+	FilterAlt = 'filter-alt',
+	FilterAltOff = 'filter-alt-off',
+	FilterBAndW = 'filter-b-and-w',
+	FilterCenterFocus = 'filter-center-focus',
+	FilterDrama = 'filter-drama',
+	FilterFrames = 'filter-frames',
+	FilterHdr = 'filter-hdr',
+	FilterList = 'filter-list',
+	FilterListOff = 'filter-list-off',
+	FilterNone = 'filter-none',
+	FilterTiltShift = 'filter-tilt-shift',
+	FilterVintage = 'filter-vintage',
+	FinanceChip = 'finance-chip',
+	FindInPage = 'find-in-page',
+	FindReplace = 'find-replace',
+	Fingerprint = 'fingerprint',
+	FireExtinguisher = 'fire-extinguisher',
+	FireHydrant = 'fire-hydrant',
+	FireTruck = 'fire-truck',
+	Fireplace = 'fireplace',
+	FirstPage = 'first-page',
+	FitScreen = 'fit-screen',
+	FitnessCenter = 'fitness-center',
+	Flag = 'flag',
+	FlagCircle = 'flag-circle',
+	Flaky = 'flaky',
+	Flare = 'flare',
+	FlashAuto = 'flash-auto',
+	FlashOff = 'flash-off',
+	FlashOn = 'flash-on',
+	FlashlightOff = 'flashlight-off',
+	FlashlightOn = 'flashlight-on',
+	Flatware = 'flatware',
+	Flight = 'flight',
+	FlightClass = 'flight-class',
+	FlightLand = 'flight-land',
+	FlightTakeoff = 'flight-takeoff',
+	Flip = 'flip',
+	FlipCameraAndroid = 'flip-camera-android',
+	FlipCameraIos = 'flip-camera-ios',
+	FlipToBack = 'flip-to-back',
+	FlipToFront = 'flip-to-front',
+	Flood = 'flood',
+	FloorLamp = 'floor-lamp',
+	Flowsheet = 'flowsheet',
+	Fluid = 'fluid',
+	FluidBalance = 'fluid-balance',
+	FluidMed = 'fluid-med',
+	Fluorescent = 'fluorescent',
+	FlutterDash = 'flutter-dash',
+	FmdBad = 'fmd-bad',
+	Foggy = 'foggy',
+	Folder = 'folder',
+	FolderCopy = 'folder-copy',
+	FolderDelete = 'folder-delete',
+	FolderOff = 'folder-off',
+	FolderOpen = 'folder-open',
+	FolderShared = 'folder-shared',
+	FolderSpecial = 'folder-special',
+	FolderZip = 'folder-zip',
+	FollowTheSigns = 'follow-the-signs',
+	FontDownload = 'font-download',
+	FontDownloadOff = 'font-download-off',
+	FoodBank = 'food-bank',
+	FootBones = 'foot-bones',
+	Footprint = 'footprint',
+	Forest = 'forest',
+	ForkLeft = 'fork-left',
+	ForkRight = 'fork-right',
+	Forklift = 'forklift',
+	FormatAlignCenter = 'format-align-center',
+	FormatAlignJustify = 'format-align-justify',
+	FormatAlignLeft = 'format-align-left',
+	FormatAlignRight = 'format-align-right',
+	FormatBold = 'format-bold',
+	FormatClear = 'format-clear',
+	FormatColorFill = 'format-color-fill',
+	FormatColorReset = 'format-color-reset',
+	FormatColorText = 'format-color-text',
+	FormatH1 = 'format-h1',
+	FormatH2 = 'format-h2',
+	FormatH3 = 'format-h3',
+	FormatH4 = 'format-h4',
+	FormatH5 = 'format-h5',
+	FormatH6 = 'format-h6',
+	FormatImageLeft = 'format-image-left',
+	FormatImageRight = 'format-image-right',
+	FormatIndentDecrease = 'format-indent-decrease',
+	FormatIndentIncrease = 'format-indent-increase',
+	FormatInkHighlighter = 'format-ink-highlighter',
+	FormatItalic = 'format-italic',
+	FormatLineSpacing = 'format-line-spacing',
+	FormatListBulleted = 'format-list-bulleted',
+	FormatListBulletedAdd = 'format-list-bulleted-add',
+	FormatListNumbered = 'format-list-numbered',
+	FormatListNumberedRtl = 'format-list-numbered-rtl',
+	FormatOverline = 'format-overline',
+	FormatPaint = 'format-paint',
+	FormatParagraph = 'format-paragraph',
+	FormatQuote = 'format-quote',
+	FormatShapes = 'format-shapes',
+	FormatSize = 'format-size',
+	FormatStrikethrough = 'format-strikethrough',
+	FormatTextClip = 'format-text-clip',
+	FormatTextOverflow = 'format-text-overflow',
+	FormatTextWrap = 'format-text-wrap',
+	FormatTextDirectionLToR = 'format-text-direction-l-to-r',
+	FormatTextDirectionRToL = 'format-text-direction-r-to-l',
+	FormatUnderlined = 'format-underlined',
+	FormatUnderlinedSquiggle = 'format-underlined-squiggle',
+	FormsAddOn = 'forms-add-on',
+	FormsAppsScript = 'forms-apps-script',
+	Fort = 'fort',
+	Forum = 'forum',
+	Forward = 'forward',
+	Forward_10 = 'forward-10',
+	Forward_30 = 'forward-30',
+	Forward_5 = 'forward-5',
+	ForwardToInbox = 'forward-to-inbox',
+	Foundation = 'foundation',
+	FramePerson = 'frame-person',
+	FreeCancellation = 'free-cancellation',
+	FrontHand = 'front-hand',
+	FrontLoader = 'front-loader',
+	FullStackedBarChart = 'full-stacked-bar-chart',
+	Fullscreen = 'fullscreen',
+	FullscreenExit = 'fullscreen-exit',
+	Function = 'function',
+	Functions = 'functions',
+	GMobileData = 'g-mobiledata',
+	GTranslate = 'g-translate',
+	GalleryThumbnail = 'gallery-thumbnail',
+	Gamepad = 'gamepad',
+	Garage = 'garage',
+	GarageHome = 'garage-home',
+	GardenCart = 'garden-cart',
+	GasMeter = 'gas-meter',
+	Gastroenterology = 'gastroenterology',
+	Gate = 'gate',
+	Gavel = 'gavel',
+	GeneratingTokens = 'generating-tokens',
+	Genetics = 'genetics',
+	Gesture = 'gesture',
+	Gif = 'gif',
+	GifBox = 'gif-box',
+	Girl = 'girl',
+	Gite = 'gite',
+	Glucose = 'glucose',
+	Glyphs = 'glyphs',
+	GolfCourse = 'golf-course',
+	GooglePlusReshare = 'google-plus-reshare',
+	GoogleWifi = 'google-wifi',
+	GppBad = 'gpp-bad',
+	GppMaybe = 'gpp-maybe',
+	Grade = 'grade',
+	Gradient = 'gradient',
+	Grading = 'grading',
+	Grain = 'grain',
+	GraphicEq = 'graphic-eq',
+	Grass = 'grass',
+	Grid_3x3 = 'grid-3x3',
+	Grid_4x4 = 'grid-4x4',
+	GridGoldenRatio = 'grid-goldenratio',
+	GridOff = 'grid-off',
+	GridOn = 'grid-on',
+	GridView = 'grid-view',
+	Group = 'group',
+	GroupAdd = 'group-add',
+	GroupOff = 'group-off',
+	GroupRemove = 'group-remove',
+	GroupWork = 'group-work',
+	GroupedBarChart = 'grouped-bar-chart',
+	Groups = 'groups',
+	Groups_2 = 'groups-2',
+	Groups_3 = 'groups-3',
+	HMobileData = 'h-mobiledata',
+	HPlusMobileData = 'h-plus-mobiledata',
+	Hail = 'hail',
+	HandBones = 'hand-bones',
+	HandGesture = 'hand-gesture',
+	Handshake = 'handshake',
+	Handyman = 'handyman',
+	HardDrive = 'hard-drive',
+	Hardware = 'hardware',
+	Hd = 'hd',
+	HdrAuto = 'hdr-auto',
+	HdrAutoSelect = 'hdr-auto-select',
+	HdrEnhancedSelect = 'hdr-enhanced-select',
+	HdrOff = 'hdr-off',
+	HdrOffSelect = 'hdr-off-select',
+	HdrOn = 'hdr-on',
+	HdrOnSelect = 'hdr-on-select',
+	HdrPlus = 'hdr-plus',
+	HdrStrong = 'hdr-strong',
+	HdrWeak = 'hdr-weak',
+	Headphones = 'headphones',
+	HeadphonesBattery = 'headphones-battery',
+	HeadsetMic = 'headset-mic',
+	HeadsetOff = 'headset-off',
+	Healing = 'healing',
+	HealthAndSafety = 'health-and-safety',
+	Hearing = 'hearing',
+	HearingDisabled = 'hearing-disabled',
+	HeartBroken = 'heart-broken',
+	HeartMinus = 'heart-minus',
+	HeartPlus = 'heart-plus',
+	HeatPump = 'heat-pump',
+	HeatPumpBalance = 'heat-pump-balance',
+	Height = 'height',
+	Help = 'help',
+	HelpCenter = 'help-center',
+	HelpClinic = 'help-clinic',
+	Hematology = 'hematology',
+	Hevc = 'hevc',
+	Hexagon = 'hexagon',
+	Hide = 'hide',
+	HideImage = 'hide-image',
+	HideSource = 'hide-source',
+	HighQuality = 'high-quality',
+	Highlight = 'highlight',
+	Hiking = 'hiking',
+	History = 'history',
+	HistoryEdu = 'history-edu',
+	HistoryToggleOff = 'history-toggle-off',
+	Hive = 'hive',
+	Hls = 'hls',
+	HlsOff = 'hls-off',
+	HolidayVillage = 'holiday-village',
+	Home = 'home',
+	HomeAppLogo = 'home-app-logo',
+	HomeHealth = 'home-health',
+	HomeIotDevice = 'home-iot-device',
+	HomeMax = 'home-max',
+	HomeMaxDots = 'home-max-dots',
+	HomeMini = 'home-mini',
+	HomePin = 'home-pin',
+	HomeRepairService = 'home-repair-service',
+	HomeSpeaker = 'home-speaker',
+	HomeStorage = 'home-storage',
+	HomeWork = 'home-work',
+	HorizontalDistribute = 'horizontal-distribute',
+	HorizontalRule = 'horizontal-rule',
+	HorizontalSplit = 'horizontal-split',
+	HotTub = 'hot-tub',
+	Hotel = 'hotel',
+	HotelClass = 'hotel-class',
+	HourglassBottom = 'hourglass-bottom',
+	HourglassDisabled = 'hourglass-disabled',
+	HourglassEmpty = 'hourglass-empty',
+	HourglassFull = 'hourglass-full',
+	HourglassTop = 'hourglass-top',
+	House = 'house',
+	HouseSiding = 'house-siding',
+	HouseWithShield = 'house-with-shield',
+	Houseboat = 'houseboat',
+	HowToReg = 'how-to-reg',
+	HowToVote = 'how-to-vote',
+	Html = 'html',
+	Http = 'http',
+	Hub = 'hub',
+	Humerus = 'humerus',
+	HumerusAlt = 'humerus-alt',
+	HumidityHigh = 'humidity-high',
+	HumidityLow = 'humidity-low',
+	HumidityMid = 'humidity-mid',
+	HumidityPercentage = 'humidity-percentage',
+	Hvac = 'hvac',
+	IceSkating = 'ice-skating',
+	Icecream = 'icecream',
+	Image = 'image',
+	ImageAspectRatio = 'image-aspect-ratio',
+	ImageNotSupported = 'image-not-supported',
+	ImageSearch = 'image-search',
+	ImagesearchRoller = 'imagesearch-roller',
+	Imagesmode = 'imagesmode',
+	Immunology = 'immunology',
+	ImportContacts = 'import-contacts',
+	ImportantDevices = 'important-devices',
+	InHomeMode = 'in-home-mode',
+	InactiveOrder = 'inactive-order',
+	Inbox = 'inbox',
+	InboxCustomize = 'inbox-customize',
+	IncompleteCircle = 'incomplete-circle',
+	IndeterminateCheckBox = 'indeterminate-check-box',
+	Info = 'info',
+	Infrared = 'infrared',
+	Inpatient = 'inpatient',
+	Input = 'input',
+	InsertChart = 'insert-chart',
+	InsertPageBreak = 'insert-page-break',
+	InsertText = 'insert-text',
+	Insights = 'insights',
+	InstallDesktop = 'install-desktop',
+	InstallMobile = 'install-mobile',
+	IntegrationInstructions = 'integration-instructions',
+	Interests = 'interests',
+	InterpreterMode = 'interpreter-mode',
+	Inventory = 'inventory',
+	Inventory_2 = 'inventory-2',
+	InvertColors = 'invert-colors',
+	InvertColorsOff = 'invert-colors-off',
+	IosShare = 'ios-share',
+	Iron = 'iron',
+	Javascript = 'javascript',
+	Join = 'join',
+	JoinFull = 'join-full',
+	JoinInner = 'join-inner',
+	JoinLeft = 'join-left',
+	JoinRight = 'join-right',
+	Kayaking = 'kayaking',
+	KebabDining = 'kebab-dining',
+	Kettle = 'kettle',
+	Key = 'key',
+	KeyOff = 'key-off',
+	KeyVisualizer = 'key-visualizer',
+	Keyboard = 'keyboard',
+	KeyboardAlt = 'keyboard-alt',
+	KeyboardArrowDown = 'keyboard-arrow-down',
+	KeyboardArrowLeft = 'keyboard-arrow-left',
+	KeyboardArrowRight = 'keyboard-arrow-right',
+	KeyboardArrowUp = 'keyboard-arrow-up',
+	KeyboardBackspace = 'keyboard-backspace',
+	KeyboardCapslock = 'keyboard-capslock',
+	KeyboardCommandKey = 'keyboard-command-key',
+	KeyboardControlKey = 'keyboard-control-key',
+	KeyboardDoubleArrowDown = 'keyboard-double-arrow-down',
+	KeyboardDoubleArrowLeft = 'keyboard-double-arrow-left',
+	KeyboardDoubleArrowRight = 'keyboard-double-arrow-right',
+	KeyboardDoubleArrowUp = 'keyboard-double-arrow-up',
+	KeyboardHide = 'keyboard-hide',
+	KeyboardOptionKey = 'keyboard-option-key',
+	KeyboardReturn = 'keyboard-return',
+	KeyboardTab = 'keyboard-tab',
+	KeyboardVoice = 'keyboard-voice',
+	KingBed = 'king-bed',
+	Kitchen = 'kitchen',
+	Kitesurfing = 'kitesurfing',
+	LabPanel = 'lab-panel',
+	LabProfile = 'lab-profile',
+	LabResearch = 'lab-research',
+	Label = 'label',
+	LabelImportant = 'label-important',
+	LabelOff = 'label-off',
+	Labs = 'labs',
+	Lan = 'lan',
+	Landscape = 'landscape',
+	Landslide = 'landslide',
+	Language = 'language',
+	LaptopChromebook = 'laptop-chromebook',
+	LaptopMac = 'laptop-mac',
+	LaptopWindows = 'laptop-windows',
+	LastPage = 'last-page',
+	Laundry = 'laundry',
+	Layers = 'layers',
+	LayersClear = 'layers-clear',
+	Lda = 'lda',
+	Leaderboard = 'leaderboard',
+	LeakAdd = 'leak-add',
+	LeakRemove = 'leak-remove',
+	LegendToggle = 'legend-toggle',
+	Lens = 'lens',
+	LensBlur = 'lens-blur',
+	LibraryAdd = 'library-add',
+	LibraryAddCheck = 'library-add-check',
+	LibraryBooks = 'library-books',
+	LibraryMusic = 'library-music',
+	LiftToTalk = 'lift-to-talk',
+	Light = 'light',
+	LightGroup = 'light-group',
+	LightMode = 'light-mode',
+	Lightbulb = 'lightbulb',
+	LightbulbCircle = 'lightbulb-circle',
+	LineAxis = 'line-axis',
+	LineEnd = 'line-end',
+	LineEndArrow = 'line-end-arrow',
+	LineEndArrowNotch = 'line-end-arrow-notch',
+	LineEndCircle = 'line-end-circle',
+	LineEndDiamond = 'line-end-diamond',
+	LineEndSquare = 'line-end-square',
+	LineStart = 'line-start',
+	LineStartArrow = 'line-start-arrow',
+	LineStartArrowNotch = 'line-start-arrow-notch',
+	LineStartCircle = 'line-start-circle',
+	LineStartDiamond = 'line-start-diamond',
+	LineStartSquare = 'line-start-square',
+	LineStyle = 'line-style',
+	LineWeight = 'line-weight',
+	LinearScale = 'linear-scale',
+	Link = 'link',
+	LinkOff = 'link-off',
+	LinkedCamera = 'linked-camera',
+	Liquor = 'liquor',
+	List = 'list',
+	ListAlt = 'list-alt',
+	LiveHelp = 'live-help',
+	LiveTv = 'live-tv',
+	Living = 'living',
+	LocalActivity = 'local-activity',
+	LocalAtm = 'local-atm',
+	LocalBar = 'local-bar',
+	LocalCafe = 'local-cafe',
+	LocalCarWash = 'local-car-wash',
+	LocalConvenienceStore = 'local-convenience-store',
+	LocalDining = 'local-dining',
+	LocalDrink = 'local-drink',
+	LocalFireDepartment = 'local-fire-department',
+	LocalFlorist = 'local-florist',
+	LocalGasStation = 'local-gas-station',
+	LocalHospital = 'local-hospital',
+	LocalLaundryService = 'local-laundry-service',
+	LocalLibrary = 'local-library',
+	LocalMall = 'local-mall',
+	LocalParking = 'local-parking',
+	LocalPharmacy = 'local-pharmacy',
+	LocalPizza = 'local-pizza',
+	LocalPolice = 'local-police',
+	LocalPostOffice = 'local-post-office',
+	LocalSee = 'local-see',
+	LocalShipping = 'local-shipping',
+	LocalTaxi = 'local-taxi',
+	LocationAutomation = 'location-automation',
+	LocationAway = 'location-away',
+	LocationChip = 'location-chip',
+	LocationCity = 'location-city',
+	LocationDisabled = 'location-disabled',
+	LocationHome = 'location-home',
+	LocationOff = 'location-off',
+	LocationOn = 'location-on',
+	LocationSearching = 'location-searching',
+	Lock = 'lock',
+	LockClock = 'lock-clock',
+	LockOpen = 'lock-open',
+	LockPerson = 'lock-person',
+	LockReset = 'lock-reset',
+	Login = 'login',
+	LogoDev = 'logo-dev',
+	Logout = 'logout',
+	Looks = 'looks',
+	Looks_3 = 'looks-3',
+	Looks_4 = 'looks-4',
+	Looks_5 = 'looks-5',
+	Looks_6 = 'looks-6',
+	LooksOne = 'looks-one',
+	LooksTwo = 'looks-two',
+	Loupe = 'loupe',
+	LowPriority = 'low-priority',
+	Loyalty = 'loyalty',
+	LteMobileData = 'lte-mobiledata',
+	LtePlusMobileData = 'lte-plus-mobiledata',
+	Luggage = 'luggage',
+	LunchDining = 'lunch-dining',
+	Lyrics = 'lyrics',
+	MacroOff = 'macro-off',
+	MagicButton = 'magic-button',
+	MagnificationLarge = 'magnification-large',
+	MagnificationSmall = 'magnification-small',
+	Mail = 'mail',
+	MailLock = 'mail-lock',
+	Male = 'male',
+	Man = 'man',
+	Man_2 = 'man-2',
+	Man_3 = 'man-3',
+	Man_4 = 'man-4',
+	ManageAccounts = 'manage-accounts',
+	ManageHistory = 'manage-history',
+	ManageSearch = 'manage-search',
+	Map = 'map',
+	MapsUgc = 'maps-ugc',
+	Margin = 'margin',
+	MarkAsUnread = 'mark-as-unread',
+	MarkChatRead = 'mark-chat-read',
+	MarkChatUnread = 'mark-chat-unread',
+	MarkEmailRead = 'mark-email-read',
+	MarkEmailUnread = 'mark-email-unread',
+	MarkUnreadChatAlt = 'mark-unread-chat-alt',
+	MarkUnreadMailbox = 'markunread-mailbox',
+	Masks = 'masks',
+	Matter = 'matter',
+	Maximize = 'maximize',
+	MediaBluetoothOff = 'media-bluetooth-off',
+	MediaBluetoothOn = 'media-bluetooth-on',
+	MediaLink = 'media-link',
+	Mediation = 'mediation',
+	MedicalInformation = 'medical-information',
+	MedicalMask = 'medical-mask',
+	MedicalServices = 'medical-services',
+	Medication = 'medication',
+	MedicationLiquid = 'medication-liquid',
+	MeetingRoom = 'meeting-room',
+	Memory = 'memory',
+	Menu = 'menu',
+	MenuBook = 'menu-book',
+	MenuOpen = 'menu-open',
+	Merge = 'merge',
+	MergeType = 'merge-type',
+	MfgNestYaleLock = 'mfg-nest-yale-lock',
+	Mic = 'mic',
+	MicExternalOff = 'mic-external-off',
+	MicExternalOn = 'mic-external-on',
+	MicOff = 'mic-off',
+	Microbiology = 'microbiology',
+	Microwave = 'microwave',
+	MicrowaveGen = 'microwave-gen',
+	MilitaryTech = 'military-tech',
+	Minimize = 'minimize',
+	MinorCrash = 'minor-crash',
+	MissedVideoCall = 'missed-video-call',
+	MixtureMed = 'mixture-med',
+	Mms = 'mms',
+	MobileFriendly = 'mobile-friendly',
+	MobileOff = 'mobile-off',
+	MobileScreenShare = 'mobile-screen-share',
+	MobileDataOff = 'MobileData-off',
+	ModeComment = 'mode-comment',
+	ModeCool = 'mode-cool',
+	ModeCoolOff = 'mode-cool-off',
+	ModeFan = 'mode-fan',
+	ModeFanOff = 'mode-fan-off',
+	ModeHeat = 'mode-heat',
+	ModeHeatCool = 'mode-heat-cool',
+	ModeHeatOff = 'mode-heat-off',
+	ModeNight = 'mode-night',
+	ModeOfTravel = 'mode-of-travel',
+	ModeOffOn = 'mode-off-on',
+	ModeStandby = 'mode-standby',
+	ModelTraining = 'model-training',
+	MonetizationOn = 'monetization-on',
+	Money = 'money',
+	MoneyOff = 'money-off',
+	Monitor = 'monitor',
+	MonitorHeart = 'monitor-heart',
+	MonitorWeight = 'monitor-weight',
+	Monitoring = 'monitoring',
+	MonochromePhotos = 'monochrome-photos',
+	Mood = 'mood',
+	MoodBad = 'mood-bad',
+	Mop = 'mop',
+	More = 'more',
+	MoreDown = 'more-down',
+	MoreHoriz = 'more-horiz',
+	MoreTime = 'more-time',
+	MoreUp = 'more-up',
+	MoreVert = 'more-vert',
+	Mosque = 'mosque',
+	MotionBlur = 'motion-blur',
+	MotionMode = 'motion-mode',
+	MotionPhotosAuto = 'motion-photos-auto',
+	MotionPhotosOff = 'motion-photos-off',
+	MotionPhotosPaused = 'motion-photos-paused',
+	MotionSensorActive = 'motion-sensor-active',
+	MotionSensorAlert = 'motion-sensor-alert',
+	MotionSensorIdle = 'motion-sensor-idle',
+	MotionSensorUrgent = 'motion-sensor-urgent',
+	Motorcycle = 'motorcycle',
+	Mouse = 'mouse',
+	MoveDown = 'move-down',
+	MoveToInbox = 'move-to-inbox',
+	MoveUp = 'move-up',
+	Movie = 'movie',
+	MovieEdit = 'movie-edit',
+	MovieFilter = 'movie-filter',
+	Moving = 'moving',
+	MovingBeds = 'moving-beds',
+	MovingMinistry = 'moving-ministry',
+	Mp = 'mp',
+	Multicooker = 'multicooker',
+	MultilineChart = 'multiline-chart',
+	MultipleStop = 'multiple-stop',
+	Museum = 'museum',
+	MusicNote = 'music-note',
+	MusicOff = 'music-off',
+	MusicVideo = 'music-video',
+	MyLocation = 'my-location',
+	Nat = 'nat',
+	Nature = 'nature',
+	NaturePeople = 'nature-people',
+	NavigateBefore = 'navigate-before',
+	NavigateNext = 'navigate-next',
+	Navigation = 'navigation',
+	NearMe = 'near-me',
+	NearMeDisabled = 'near-me-disabled',
+	NearbyError = 'nearby-error',
+	NearbyOff = 'nearby-off',
+	Nephrology = 'nephrology',
+	NestAudio = 'nest-audio',
+	NestCamFloodlight = 'nest-cam-floodlight',
+	NestCamIndoor = 'nest-cam-indoor',
+	NestCamIq = 'nest-cam-iq',
+	NestCamIqOutdoor = 'nest-cam-iq-outdoor',
+	NestCamMagnetMount = 'nest-cam-magnet-mount',
+	NestCamOutdoor = 'nest-cam-outdoor',
+	NestCamStand = 'nest-cam-stand',
+	NestCamWallMount = 'nest-cam-wall-mount',
+	NestCamWiredStand = 'nest-cam-wired-stand',
+	NestClockFarsightAnalog = 'nest-clock-farsight-analog',
+	NestClockFarsightDigital = 'nest-clock-farsight-digital',
+	NestConnect = 'nest-connect',
+	NestDetect = 'nest-detect',
+	NestDisplay = 'nest-display',
+	NestDisplayMax = 'nest-display-max',
+	NestDoorbellVisitor = 'nest-doorbell-visitor',
+	NestEcoLeaf = 'nest-eco-leaf',
+	NestFarsightWeather = 'nest-farsight-weather',
+	NestFoundSavings = 'nest-found-savings',
+	NestHeatLinkE = 'nest-heat-link-e',
+	NestHeatLinkGen_3 = 'nest-heat-link-gen-3',
+	NestHelloDoorbell = 'nest-hello-doorbell',
+	NestLocatorTag = 'nest-locator-tag',
+	NestMini = 'nest-mini',
+	NestMultiRoom = 'nest-multi-room',
+	NestRemote = 'nest-remote',
+	NestRemoteComfortSensor = 'nest-remote-comfort-sensor',
+	NestSecureAlarm = 'nest-secure-alarm',
+	NestSunblock = 'nest-sunblock',
+	NestTag = 'nest-tag',
+	NestThermostatEEu = 'nest-thermostat-e-eu',
+	NestThermostatGen_3 = 'nest-thermostat-gen-3',
+	NestThermostatSensor = 'nest-thermostat-sensor',
+	NestThermostatSensorEu = 'nest-thermostat-sensor-eu',
+	NestThermostatZirconiumEu = 'nest-thermostat-zirconium-eu',
+	NestTrueRadiant = 'nest-true-radiant',
+	NestWakeOnApproach = 'nest-wake-on-approach',
+	NestWakeOnPress = 'nest-wake-on-press',
+	NestWifiGale = 'nest-wifi-gale',
+	NestWifiMistral = 'nest-wifi-mistral',
+	NestWifiPoint = 'nest-wifi-point',
+	NestWifiPointVento = 'nest-wifi-point-vento',
+	NestWifiRouter = 'nest-wifi-router',
+	NetworkCell = 'network-cell',
+	NetworkCheck = 'network-check',
+	NetworkLocked = 'network-locked',
+	NetworkPing = 'network-ping',
+	NetworkWifi = 'network-wifi',
+	NetworkWifi_1Bar = 'network-wifi-1-bar',
+	NetworkWifi_2Bar = 'network-wifi-2-bar',
+	NetworkWifi_3Bar = 'network-wifi-3-bar',
+	NewLabel = 'new-label',
+	NewReleases = 'new-releases',
+	Newspaper = 'newspaper',
+	NextPlan = 'next-plan',
+	NextWeek = 'next-week',
+	Nfc = 'nfc',
+	NightShelter = 'night-shelter',
+	NightSightAuto = 'night-sight-auto',
+	NightSightAutoOff = 'night-sight-auto-off',
+	Nightlife = 'nightlife',
+	Nightlight = 'nightlight',
+	NightsStay = 'nights-stay',
+	NoAccounts = 'no-accounts',
+	NoAdultContent = 'no-adult-content',
+	NoBackpack = 'no-backpack',
+	NoCrash = 'no-crash',
+	NoDrinks = 'no-drinks',
+	NoEncryption = 'no-encryption',
+	NoFlash = 'no-flash',
+	NoFood = 'no-food',
+	NoLuggage = 'no-luggage',
+	NoMeals = 'no-meals',
+	NoMeetingRoom = 'no-meeting-room',
+	NoPhotography = 'no-photography',
+	NoSim = 'no-sim',
+	NoStroller = 'no-stroller',
+	NoTransfer = 'no-transfer',
+	NoiseAware = 'noise-aware',
+	NoiseControlOff = 'noise-control-off',
+	NoiseControlOn = 'noise-control-on',
+	NordicWalking = 'nordic-walking',
+	North = 'north',
+	NorthEast = 'north-east',
+	NorthWest = 'north-west',
+	NotAccessible = 'not-accessible',
+	NotListedLocation = 'not-listed-location',
+	NotStarted = 'not-started',
+	Note = 'note',
+	NoteAdd = 'note-add',
+	NoteAlt = 'note-alt',
+	Notes = 'notes',
+	NotificationAdd = 'notification-add',
+	NotificationImportant = 'notification-important',
+	Notifications = 'notifications',
+	NotificationsActive = 'notifications-active',
+	NotificationsOff = 'notifications-off',
+	NotificationsPaused = 'notifications-paused',
+	Numbers = 'numbers',
+	Nutrition = 'nutrition',
+	OfflineBolt = 'offline-bolt',
+	OfflinePin = 'offline-pin',
+	OfflineShare = 'offline-share',
+	OilBarrel = 'oil-barrel',
+	OnDeviceTraining = 'on-device-training',
+	Oncology = 'oncology',
+	OnlinePrediction = 'online-prediction',
+	Opacity = 'opacity',
+	OpenInBrowser = 'open-in-browser',
+	OpenInFull = 'open-in-full',
+	OpenInNew = 'open-in-new',
+	OpenInNewOff = 'open-in-new-off',
+	OpenWith = 'open-with',
+	Ophthalmology = 'ophthalmology',
+	OralDisease = 'oral-disease',
+	OrderApprove = 'order-approve',
+	OrderPlay = 'order-play',
+	Orthopedics = 'orthopedics',
+	OtherAdmission = 'other-admission',
+	OtherHouses = 'other-houses',
+	Outbound = 'outbound',
+	Outbox = 'outbox',
+	OutdoorGarden = 'outdoor-garden',
+	OutdoorGrill = 'outdoor-grill',
+	OutgoingMail = 'outgoing-mail',
+	Outlet = 'outlet',
+	Outpatient = 'outpatient',
+	OutpatientMed = 'outpatient-med',
+	Output = 'output',
+	OvenGen = 'oven-gen',
+	Overview = 'overview',
+	Pacemaker = 'pacemaker',
+	Package = 'package',
+	Padding = 'padding',
+	Pages = 'pages',
+	PageView = 'pageview',
+	Paid = 'paid',
+	Palette = 'palette',
+	Pallet = 'pallet',
+	PanTool = 'pan-tool',
+	PanToolAlt = 'pan-tool-alt',
+	Panorama = 'panorama',
+	PanoramaFishEye = 'panorama-fish-eye',
+	PanoramaHorizontal = 'panorama-horizontal',
+	PanoramaPhotosphere = 'panorama-photosphere',
+	PanoramaVertical = 'panorama-vertical',
+	PanoramaWideAngle = 'panorama-wide-angle',
+	Paragliding = 'paragliding',
+	Park = 'park',
+	PartlyCloudyDay = 'partly-cloudy-day',
+	PartlyCloudyNight = 'partly-cloudy-night',
+	PartyMode = 'party-mode',
+	Password = 'password',
+	PatientList = 'patient-list',
+	Pattern = 'pattern',
+	Pause = 'pause',
+	PauseCircle = 'pause-circle',
+	PausePresentation = 'pause-presentation',
+	Payments = 'payments',
+	PedalBike = 'pedal-bike',
+	Pediatrics = 'pediatrics',
+	Pending = 'pending',
+	PendingActions = 'pending-actions',
+	Pentagon = 'pentagon',
+	Percent = 'percent',
+	Pergola = 'pergola',
+	PermCameraMic = 'perm-camera-mic',
+	PermContactCalendar = 'perm-contact-calendar',
+	PermDataSetting = 'perm-data-setting',
+	PermDeviceInformation = 'perm-device-information',
+	PermMedia = 'perm-media',
+	PermPhoneMsg = 'perm-phone-msg',
+	PermScanWifi = 'perm-scan-wifi',
+	Person = 'person',
+	Person_2 = 'person-2',
+	Person_3 = 'person-3',
+	Person_4 = 'person-4',
+	PersonAdd = 'person-add',
+	PersonAddDisabled = 'person-add-disabled',
+	PersonFilled = 'person-filled',
+	PersonOff = 'person-off',
+	PersonPin = 'person-pin',
+	PersonPinCircle = 'person-pin-circle',
+	PersonRemove = 'person-remove',
+	PersonSearch = 'person-search',
+	PersonalInjury = 'personal-injury',
+	PestControl = 'pest-control',
+	PestControlRodent = 'pest-control-rodent',
+	Pets = 'pets',
+	Phishing = 'phishing',
+	PhoneAndroid = 'phone-android',
+	PhoneBluetoothSpeaker = 'phone-bluetooth-speaker',
+	PhoneCallback = 'phone-callback',
+	PhoneDisabled = 'phone-disabled',
+	PhoneEnabled = 'phone-enabled',
+	PhoneForwarded = 'phone-forwarded',
+	PhoneInTalk = 'phone-in-talk',
+	PhoneInTalkWatchfaceIndicator = 'phone-in-talk-watchface-indicator',
+	PhoneIphone = 'phone-iphone',
+	PhoneLocked = 'phone-locked',
+	PhoneMissed = 'phone-missed',
+	PhonePaused = 'phone-paused',
+	PhonelinkErase = 'phonelink-erase',
+	PhonelinkLock = 'phonelink-lock',
+	PhonelinkOff = 'phonelink-off',
+	PhonelinkRing = 'phonelink-ring',
+	PhonelinkSetup = 'phonelink-setup',
+	Photo = 'photo',
+	PhotoAlbum = 'photo-album',
+	PhotoCamera = 'photo-camera',
+	PhotoCameraBack = 'photo-camera-back',
+	PhotoCameraFront = 'photo-camera-front',
+	PhotoFilter = 'photo-filter',
+	PhotoFrame = 'photo-frame',
+	PhotoLibrary = 'photo-library',
+	PhotoSizeSelectLarge = 'photo-size-select-large',
+	PhotoSizeSelectSmall = 'photo-size-select-small',
+	Php = 'php',
+	Piano = 'piano',
+	PianoOff = 'piano-off',
+	PictureAsPdf = 'picture-as-pdf',
+	PictureInPicture = 'picture-in-picture',
+	PictureInPictureAlt = 'picture-in-picture-alt',
+	PieChart = 'pie-chart',
+	Pin = 'pin',
+	PinDrop = 'pin-drop',
+	PinEnd = 'pin-end',
+	PinInvoke = 'pin-invoke',
+	Pinch = 'pinch',
+	PinchZoomIn = 'pinch-zoom-in',
+	PinchZoomOut = 'pinch-zoom-out',
+	PivotTableChart = 'pivot-table-chart',
+	PlaceItem = 'place-item',
+	Plagiarism = 'plagiarism',
+	PlayArrow = 'play-arrow',
+	PlayCircle = 'play-circle',
+	PlayDisabled = 'play-disabled',
+	PlayForWork = 'play-for-work',
+	PlayLesson = 'play-lesson',
+	PlayPause = 'play-pause',
+	PlaylistAdd = 'playlist-add',
+	PlaylistAddCheck = 'playlist-add-check',
+	PlaylistAddCheckCircle = 'playlist-add-check-circle',
+	PlaylistAddCircle = 'playlist-add-circle',
+	PlaylistPlay = 'playlist-play',
+	PlaylistRemove = 'playlist-remove',
+	Plumbing = 'plumbing',
+	Podcasts = 'podcasts',
+	Podiatry = 'podiatry',
+	PointOfSale = 'point-of-sale',
+	Policy = 'policy',
+	Polyline = 'polyline',
+	Polymer = 'polymer',
+	Pool = 'pool',
+	PortableWifiOff = 'portable-wifi-off',
+	PostAdd = 'post-add',
+	PottedPlant = 'potted-plant',
+	Power = 'power',
+	PowerInput = 'power-input',
+	PowerOff = 'power-off',
+	PowerRounded = 'power-rounded',
+	PowerSettingsNew = 'power-settings-new',
+	PrayerTimes = 'prayer-times',
+	PrecisionManufacturing = 'precision-manufacturing',
+	PregnantWoman = 'pregnant-woman',
+	Preliminary = 'preliminary',
+	Prescriptions = 'prescriptions',
+	PresentToAll = 'present-to-all',
+	Preview = 'preview',
+	PriceChange = 'price-change',
+	PriceCheck = 'price-check',
+	Print = 'print',
+	PrintDisabled = 'print-disabled',
+	Priority = 'priority',
+	PriorityHigh = 'priority-high',
+	Privacy = 'privacy',
+	PrivacyTip = 'privacy-tip',
+	PrivateConnectivity = 'private-connectivity',
+	Problem = 'problem',
+	ProcessChart = 'process-chart',
+	ProductionQuantityLimits = 'production-quantity-limits',
+	Productivity = 'productivity',
+	Propane = 'propane',
+	PropaneTank = 'propane-tank',
+	Psychiatry = 'psychiatry',
+	Psychology = 'psychology',
+	PsychologyAlt = 'psychology-alt',
+	Public = 'public',
+	PublicOff = 'public-off',
+	Publish = 'publish',
+	PublishedWithChanges = 'published-with-changes',
+	Pulmonology = 'pulmonology',
+	PunchClock = 'punch-clock',
+	PushPin = 'push-pin',
+	QrCode = 'qr-code',
+	QrCode_2 = 'qr-code-2',
+	QrCodeScanner = 'qr-code-scanner',
+	QueryStats = 'query-stats',
+	QuestionMark = 'question-mark',
+	QueueMusic = 'queue-music',
+	QueuePlayNext = 'queue-play-next',
+	QuickPhrases = 'quick-phrases',
+	QuickReference = 'quick-reference',
+	QuickReferenceAll = 'quick-reference-all',
+	Quickreply = 'quickreply',
+	QuietTime = 'quiet-time',
+	QuietTimeActive = 'quiet-time-active',
+	Quiz = 'quiz',
+	RMobileData = 'r-mobiledata',
+	Radar = 'radar',
+	Radio = 'radio',
+	RadioButtonChecked = 'radio-button-checked',
+	RadioButtonUnchecked = 'radio-button-unchecked',
+	Radiology = 'radiology',
+	RailwayAlert = 'railway-alert',
+	Rainy = 'rainy',
+	RamenDining = 'ramen-dining',
+	RampLeft = 'ramp-left',
+	RampRight = 'ramp-right',
+	RangeHood = 'range-hood',
+	RateReview = 'rate-review',
+	RawOff = 'raw-off',
+	RawOn = 'raw-on',
+	ReadMore = 'read-more',
+	RealEstateAgent = 'real-estate-agent',
+	Rebase = 'rebase',
+	RebaseEdit = 'rebase-edit',
+	Receipt = 'receipt',
+	ReceiptLong = 'receipt-long',
+	RecentActors = 'recent-actors',
+	RecentPatient = 'recent-patient',
+	Recommend = 'recommend',
+	RecordVoiceOver = 'record-voice-over',
+	Rectangle = 'rectangle',
+	Recycling = 'recycling',
+	Redeem = 'redeem',
+	Redo = 'redo',
+	ReduceCapacity = 'reduce-capacity',
+	Refresh = 'refresh',
+	RememberMe = 'remember-me',
+	RemoteGen = 'remote-gen',
+	Remove = 'remove',
+	RemoveDone = 'remove-done',
+	RemoveFromQueue = 'remove-from-queue',
+	RemoveModerator = 'remove-moderator',
+	RemoveRoad = 'remove-road',
+	RemoveShoppingCart = 'remove-shopping-cart',
+	Reorder = 'reorder',
+	Repartition = 'repartition',
+	Repeat = 'repeat',
+	RepeatOn = 'repeat-on',
+	RepeatOne = 'repeat-one',
+	RepeatOneOn = 'repeat-one-on',
+	Replay = 'replay',
+	Replay_10 = 'replay-10',
+	Replay_30 = 'replay-30',
+	Replay_5 = 'replay-5',
+	ReplayCircleFilled = 'replay-circle-filled',
+	Reply = 'reply',
+	ReplyAll = 'reply-all',
+	Report = 'report',
+	ReportOff = 'report-off',
+	RequestPage = 'request-page',
+	RequestQuote = 'request-quote',
+	ResetImage = 'reset-image',
+	ResetTv = 'reset-tv',
+	RestartAlt = 'restart-alt',
+	Restaurant = 'restaurant',
+	RestaurantMenu = 'restaurant-menu',
+	RestoreFromTrash = 'restore-from-trash',
+	RestorePage = 'restore-page',
+	Reviews = 'reviews',
+	Rheumatology = 'rheumatology',
+	RibCage = 'rib-cage',
+	RiceBowl = 'rice-bowl',
+	RingVolume = 'ring-volume',
+	Robot = 'robot',
+	Rocket = 'rocket',
+	RocketLaunch = 'rocket-launch',
+	RollerShades = 'roller-shades',
+	RollerShadesClosed = 'roller-shades-closed',
+	RollerSkating = 'roller-skating',
+	Roofing = 'roofing',
+	RoomPreferences = 'room-preferences',
+	RoomService = 'room-service',
+	Rotate_90DegreesCcw = 'rotate-90-degrees-ccw',
+	Rotate_90DegreesCw = 'rotate-90-degrees-cw',
+	RotateLeft = 'rotate-left',
+	RotateRight = 'rotate-right',
+	RoundaboutLeft = 'roundabout-left',
+	RoundaboutRight = 'roundabout-right',
+	RoundedCorner = 'rounded-corner',
+	Route = 'route',
+	Router = 'router',
+	Routine = 'routine',
+	Rowing = 'rowing',
+	RssFeed = 'rss-feed',
+	Rsvp = 'rsvp',
+	Rtt = 'rtt',
+	Rule = 'rule',
+	RuleFolder = 'rule-folder',
+	RunCircle = 'run-circle',
+	RunningWithErrors = 'running-with-errors',
+	RvHookup = 'rv-hookup',
+	SafetyCheck = 'safety-check',
+	SafetyDivider = 'safety-divider',
+	Sailing = 'sailing',
+	Salinity = 'salinity',
+	Sanitizer = 'sanitizer',
+	Satellite = 'satellite',
+	SatelliteAlt = 'satellite-alt',
+	Save = 'save',
+	SaveAs = 'save-as',
+	SavedSearch = 'saved-search',
+	Savings = 'savings',
+	Scale = 'scale',
+	Scanner = 'scanner',
+	ScatterPlot = 'scatter-plot',
+	Scene = 'scene',
+	Schedule = 'schedule',
+	ScheduleSend = 'schedule-send',
+	Schema = 'schema',
+	School = 'school',
+	Science = 'science',
+	Score = 'score',
+	Scoreboard = 'scoreboard',
+	ScreenLockLandscape = 'screen-lock-landscape',
+	ScreenLockPortrait = 'screen-lock-portrait',
+	ScreenLockRotation = 'screen-lock-rotation',
+	ScreenRotation = 'screen-rotation',
+	ScreenRotationAlt = 'screen-rotation-alt',
+	ScreenSearchDesktop = 'screen-search-desktop',
+	ScreenShare = 'screen-share',
+	Screenshot = 'screenshot',
+	ScreenshotMonitor = 'screenshot-monitor',
+	ScubaDiving = 'scuba-diving',
+	Sd = 'sd',
+	SdCard = 'sd-card',
+	SdCardAlert = 'sd-card-alert',
+	Search = 'search',
+	SearchCheck = 'search-check',
+	SearchOff = 'search-off',
+	Security = 'security',
+	SecurityUpdateGood = 'security-update-good',
+	SecurityUpdateWarning = 'security-update-warning',
+	Segment = 'segment',
+	SelectAll = 'select-all',
+	SelectCheckBox = 'select-check-box',
+	SelfCare = 'self-care',
+	SelfImprovement = 'self-improvement',
+	Sell = 'sell',
+	Send = 'send',
+	SendAndArchive = 'send-and-archive',
+	SendTimeExtension = 'send-time-extension',
+	SendToMobile = 'send-to-mobile',
+	SensorDoor = 'sensor-door',
+	SensorOccupied = 'sensor-occupied',
+	SensorWindow = 'sensor-window',
+	Sensors = 'sensors',
+	SensorsOff = 'sensors-off',
+	SentimentDissatisfied = 'sentiment-dissatisfied',
+	SentimentExtremelyDissatisfied = 'sentiment-extremely-dissatisfied',
+	SentimentNeutral = 'sentiment-neutral',
+	SentimentSatisfied = 'sentiment-satisfied',
+	SentimentVeryDissatisfied = 'sentiment-very-dissatisfied',
+	SentimentVerySatisfied = 'sentiment-very-satisfied',
+	SetMeal = 'set-meal',
+	Settings = 'settings',
+	SettingsAccessibility = 'settings-accessibility',
+	SettingsAccountBox = 'settings-account-box',
+	SettingsAlert = 'settings-alert',
+	SettingsApplications = 'settings-applications',
+	SettingsBackupRestore = 'settings-backup-restore',
+	SettingsBluetooth = 'settings-bluetooth',
+	SettingsBrightness = 'settings-brightness',
+	SettingsCell = 'settings-cell',
+	SettingsEthernet = 'settings-ethernet',
+	SettingsInputAntenna = 'settings-input-antenna',
+	SettingsInputComponent = 'settings-input-component',
+	SettingsInputHdmi = 'settings-input-hdmi',
+	SettingsInputSvideo = 'settings-input-svideo',
+	SettingsMotionMode = 'settings-motion-mode',
+	SettingsNightSight = 'settings-night-sight',
+	SettingsOverscan = 'settings-overscan',
+	SettingsPanorama = 'settings-panorama',
+	SettingsPhone = 'settings-phone',
+	SettingsPhotoCamera = 'settings-photo-camera',
+	SettingsPower = 'settings-power',
+	SettingsRemote = 'settings-remote',
+	SettingsSuggest = 'settings-suggest',
+	SettingsSystemDaydream = 'settings-system-daydream',
+	SettingsVoice = 'settings-voice',
+	SettopComponent = 'settop-component',
+	SevereCold = 'severe-cold',
+	ShapeLine = 'shape-line',
+	Shapes = 'shapes',
+	Share = 'share',
+	ShareLocation = 'share-location',
+	ShareReviews = 'share-reviews',
+	SheetsRtl = 'sheets-rtl',
+	Shelves = 'shelves',
+	Shield = 'shield',
+	ShieldMoon = 'shield-moon',
+	ShieldWithHeart = 'shield-with-heart',
+	ShieldWithHouse = 'shield-with-house',
+	Shop = 'shop',
+	ShopTwo = 'shop-two',
+	ShoppingBag = 'shopping-bag',
+	ShoppingBasket = 'shopping-basket',
+	ShoppingCart = 'shopping-cart',
+	ShoppingCartCheckout = 'shopping-cart-checkout',
+	ShortStay = 'short-stay',
+	ShortText = 'short-text',
+	ShowChart = 'show-chart',
+	Shower = 'shower',
+	Shuffle = 'shuffle',
+	ShuffleOn = 'shuffle-on',
+	ShutterSpeed = 'shutter-speed',
+	Sick = 'sick',
+	SignLanguage = 'sign-language',
+	SignalCellular_0Bar = 'signal-cellular-0-bar',
+	SignalCellular_1Bar = 'signal-cellular-1-bar',
+	SignalCellular_2Bar = 'signal-cellular-2-bar',
+	SignalCellular_3Bar = 'signal-cellular-3-bar',
+	SignalCellular_4Bar = 'signal-cellular-4-bar',
+	SignalCellularAlt = 'signal-cellular-alt',
+	SignalCellularAlt_1Bar = 'signal-cellular-alt-1-bar',
+	SignalCellularAlt_2Bar = 'signal-cellular-alt-2-bar',
+	SignalCellularConnectedNoInternet_0Bar = 'signal-cellular-connected-no-internet-0-bar',
+	SignalCellularConnectedNoInternet_4Bar = 'signal-cellular-connected-no-internet-4-bar',
+	SignalCellularNodata = 'signal-cellular-nodata',
+	SignalCellularNull = 'signal-cellular-null',
+	SignalCellularOff = 'signal-cellular-off',
+	SignalDisconnected = 'signal-disconnected',
+	SignalWifi_0Bar = 'signal-wifi-0-bar',
+	SignalWifi_4Bar = 'signal-wifi-4-bar',
+	SignalWifiBad = 'signal-wifi-bad',
+	SignalWifiOff = 'signal-wifi-off',
+	SignalWifiStatusbarNotConnected = 'signal-wifi-statusbar-not-connected',
+	SignalWifiStatusbarNull = 'signal-wifi-statusbar-null',
+	Signpost = 'signpost',
+	SimCard = 'sim-card',
+	SimCardDownload = 'sim-card-download',
+	SingleBed = 'single-bed',
+	Sip = 'sip',
+	Skateboarding = 'skateboarding',
+	Skeleton = 'skeleton',
+	SkipNext = 'skip-next',
+	SkipPrevious = 'skip-previous',
+	Skull = 'skull',
+	Sledding = 'sledding',
+	SlideLibrary = 'slide-library',
+	Slideshow = 'slideshow',
+	SlowMotionVideo = 'slow-motion-video',
+	SmartButton = 'smart-button',
+	SmartDisplay = 'smart-display',
+	SmartOutlet = 'smart-outlet',
+	SmartScreen = 'smart-screen',
+	SmartToy = 'smart-toy',
+	Smartphone = 'smartphone',
+	SmokeFree = 'smoke-free',
+	SmokingRooms = 'smoking-rooms',
+	Sms = 'sms',
+	SmsFailed = 'sms-failed',
+	SnippetFolder = 'snippet-folder',
+	Snooze = 'snooze',
+	Snowboarding = 'snowboarding',
+	Snowing = 'snowing',
+	Snowmobile = 'snowmobile',
+	Snowshoeing = 'snowshoeing',
+	Soap = 'soap',
+	SocialDistance = 'social-distance',
+	SolarPower = 'solar-power',
+	Sort = 'sort',
+	SortByAlpha = 'sort-by-alpha',
+	Sos = 'sos',
+	SoundDetectionDogBarking = 'sound-detection-dog-barking',
+	SoundDetectionGlassBreak = 'sound-detection-glass-break',
+	SoundDetectionLoudSound = 'sound-detection-loud-sound',
+	SoupKitchen = 'soup-kitchen',
+	SourceNotes = 'source-notes',
+	South = 'south',
+	SouthAmerica = 'south-america',
+	SouthEast = 'south-east',
+	SouthWest = 'south-west',
+	Spa = 'spa',
+	SpaceBar = 'space-bar',
+	SpaceDashboard = 'space-dashboard',
+	SpatialAudio = 'spatial-audio',
+	SpatialAudioOff = 'spatial-audio-off',
+	SpatialTracking = 'spatial-tracking',
+	Speaker = 'speaker',
+	SpeakerGroup = 'speaker-group',
+	SpeakerNotes = 'speaker-notes',
+	SpeakerNotesOff = 'speaker-notes-off',
+	SpeakerPhone = 'speaker-phone',
+	SpecificGravity = 'specific-gravity',
+	SpeechToText = 'speech-to-text',
+	Speed = 'speed',
+	Spellcheck = 'spellcheck',
+	SplitScreen = 'splitscreen',
+	Spoke = 'spoke',
+	Sports = 'sports',
+	SportsBar = 'sports-bar',
+	SportsBaseball = 'sports-baseball',
+	SportsBasketball = 'sports-basketball',
+	SportsCricket = 'sports-cricket',
+	SportsEsports = 'sports-esports',
+	SportsFootball = 'sports-football',
+	SportsGolf = 'sports-golf',
+	SportsGymnastics = 'sports-gymnastics',
+	SportsHandball = 'sports-handball',
+	SportsHockey = 'sports-hockey',
+	SportsKabaddi = 'sports-kabaddi',
+	SportsMartialArts = 'sports-martial-arts',
+	SportsMma = 'sports-mma',
+	SportsMotorsports = 'sports-motorsports',
+	SportsRugby = 'sports-rugby',
+	SportsScore = 'sports-score',
+	SportsSoccer = 'sports-soccer',
+	SportsTennis = 'sports-tennis',
+	SportsVolleyball = 'sports-volleyball',
+	Sprinkler = 'sprinkler',
+	Sprint = 'sprint',
+	Square = 'square',
+	SquareFoot = 'square-foot',
+	SsidChart = 'ssid-chart',
+	StackedBarChart = 'stacked-bar-chart',
+	StackedLineChart = 'stacked-line-chart',
+	StadiaController = 'stadia-controller',
+	Stadium = 'stadium',
+	Stairs = 'stairs',
+	Star = 'star',
+	StarHalf = 'star-half',
+	StarRate = 'star-rate',
+	StarRateHalf = 'star-rate-half',
+	Stars = 'stars',
+	Start = 'start',
+	StayCurrentLandscape = 'stay-current-landscape',
+	StayCurrentPortrait = 'stay-current-portrait',
+	StayPrimaryLandscape = 'stay-primary-landscape',
+	StayPrimaryPortrait = 'stay-primary-portrait',
+	Stethoscope = 'stethoscope',
+	StethoscopeArrow = 'stethoscope-arrow',
+	StethoscopeCheck = 'stethoscope-check',
+	StickyNote_2 = 'sticky-note-2',
+	Stop = 'stop',
+	StopCircle = 'stop-circle',
+	StopScreenShare = 'stop-screen-share',
+	Storage = 'storage',
+	Store = 'store',
+	Storefront = 'storefront',
+	Storm = 'storm',
+	Straight = 'straight',
+	Straighten = 'straighten',
+	Stream = 'stream',
+	StreetView = 'streetview',
+	StrikethroughS = 'strikethrough-s',
+	Stroller = 'stroller',
+	Style = 'style',
+	Styler = 'styler',
+	SubdirectoryArrowLeft = 'subdirectory-arrow-left',
+	SubdirectoryArrowRight = 'subdirectory-arrow-right',
+	Subject = 'subject',
+	Subscript = 'subscript',
+	Subscriptions = 'subscriptions',
+	Subtitles = 'subtitles',
+	SubtitlesOff = 'subtitles-off',
+	Subway = 'subway',
+	Summarize = 'summarize',
+	Sunny = 'sunny',
+	SunnySnowing = 'sunny-snowing',
+	Superscript = 'superscript',
+	SupervisedUserCircle = 'supervised-user-circle',
+	SupervisorAccount = 'supervisor-account',
+	Support = 'support',
+	SupportAgent = 'support-agent',
+	Surfing = 'surfing',
+	SurroundSound = 'surround-sound',
+	SwapCalls = 'swap-calls',
+	SwapHoriz = 'swap-horiz',
+	SwapHorizontalCircle = 'swap-horizontal-circle',
+	SwapVert = 'swap-vert',
+	SwapVerticalCircle = 'swap-vertical-circle',
+	Swipe = 'swipe',
+	SwipeDown = 'swipe-down',
+	SwipeDownAlt = 'swipe-down-alt',
+	SwipeLeft = 'swipe-left',
+	SwipeLeftAlt = 'swipe-left-alt',
+	SwipeRight = 'swipe-right',
+	SwipeRightAlt = 'swipe-right-alt',
+	SwipeUp = 'swipe-up',
+	SwipeUpAlt = 'swipe-up-alt',
+	SwipeVertical = 'swipe-vertical',
+	Switch = 'switch',
+	SwitchAccessShortcut = 'switch-access-shortcut',
+	SwitchAccessShortcutAdd = 'switch-access-shortcut-add',
+	SwitchAccount = 'switch-account',
+	SwitchCamera = 'switch-camera',
+	SwitchLeft = 'switch-left',
+	SwitchRight = 'switch-right',
+	SwitchVideo = 'switch-video',
+	Swords = 'swords',
+	Synagogue = 'synagogue',
+	Sync = 'sync',
+	SyncAlt = 'sync-alt',
+	SyncDisabled = 'sync-disabled',
+	SyncLock = 'sync-lock',
+	SyncProblem = 'sync-problem',
+	SyncSavedLocally = 'sync-saved-locally',
+	Syringe = 'syringe',
+	SystemUpdate = 'system-update',
+	SystemUpdateAlt = 'system-update-alt',
+	Tab = 'tab',
+	TabUnselected = 'tab-unselected',
+	Table = 'table',
+	TableBar = 'table-bar',
+	TableChart = 'table-chart',
+	TableLamp = 'table-lamp',
+	TableRestaurant = 'table-restaurant',
+	TableRows = 'table-rows',
+	TableView = 'table-view',
+	Tablet = 'tablet',
+	TabletAndroid = 'tablet-android',
+	TabletMac = 'tablet-mac',
+	Tag = 'tag',
+	TakeoutDining = 'takeout-dining',
+	TamperDetectionOff = 'tamper-detection-off',
+	TamperDetectionOn = 'tamper-detection-on',
+	TapAndPlay = 'tap-and-play',
+	Tapas = 'tapas',
+	Task = 'task',
+	TaskAlt = 'task-alt',
+	TaxiAlert = 'taxi-alert',
+	TeamDashboard = 'team-dashboard',
+	TempPreferencesCustom = 'temp-preferences-custom',
+	TempPreferencesEco = 'temp-preferences-eco',
+	TempleBuddhist = 'temple-buddhist',
+	TempleHindu = 'temple-hindu',
+	Tenancy = 'tenancy',
+	Terminal = 'terminal',
+	TextDecrease = 'text-decrease',
+	TextFields = 'text-fields',
+	TextFormat = 'text-format',
+	TextIncrease = 'text-increase',
+	TextRotateUp = 'text-rotate-up',
+	TextRotateVertical = 'text-rotate-vertical',
+	TextRotationAngleDown = 'text-rotation-angledown',
+	TextRotationAngleUp = 'text-rotation-angleup',
+	TextRotationDown = 'text-rotation-down',
+	TextRotationNone = 'text-rotation-none',
+	TextSnippet = 'text-snippet',
+	TextToSpeech = 'text-to-speech',
+	Texture = 'texture',
+	TheaterComedy = 'theater-comedy',
+	Theaters = 'theaters',
+	Thermometer = 'thermometer',
+	Thermostat = 'thermostat',
+	ThermostatAuto = 'thermostat-auto',
+	ThermostatCarbon = 'thermostat-carbon',
+	ThumbDown = 'thumb-down',
+	ThumbDownOff = 'thumb-down-off',
+	ThumbUp = 'thumb-up',
+	ThumbUpOff = 'thumb-up-off',
+	ThumbsUpDown = 'thumbs-up-down',
+	Thunderstorm = 'thunderstorm',
+	Tibia = 'tibia',
+	TibiaAlt = 'tibia-alt',
+	TimeAuto = 'time-auto',
+	Timelapse = 'timelapse',
+	TimeLapse = Timelapse,
+	Timeline = 'timeline',
+	Timer = 'timer',
+	Timer_10 = 'timer-10',
+	Timer_10Alt_1 = 'timer-10-alt-1',
+	Timer_10Select = 'timer-10-select',
+	Timer_3 = 'timer-3',
+	Timer_3Alt_1 = 'timer-3-alt-1',
+	Timer_3Select = 'timer-3-select',
+	TimerOff = 'timer-off',
+	TipsAndUpdates = 'tips-and-updates',
+	TireRepair = 'tire-repair',
+	Title = 'title',
+	Toc = 'toc',
+	Today = 'today',
+	ToggleOff = 'toggle-off',
+	ToggleOn = 'toggle-on',
+	Token = 'token',
+	Toll = 'toll',
+	Tonality = 'tonality',
+	ToolsFlatHead = 'tools-flat-head',
+	ToolsInstallationKit = 'tools-installation-kit',
+	ToolsLadder = 'tools-ladder',
+	ToolsLevel = 'tools-level',
+	ToolsPhillips = 'tools-phillips',
+	ToolsPliersWireStripper = 'tools-pliers-wire-stripper',
+	ToolsPowerDrill = 'tools-power-drill',
+	ToolsWrench = 'tools-wrench',
+	Topic = 'topic',
+	Tornado = 'tornado',
+	TotalDissolvedSolids = 'total-dissolved-solids',
+	TouchApp = 'touch-app',
+	Tour = 'tour',
+	Toys = 'toys',
+	ToysFan = 'toys-fan',
+	TrackChanges = 'track-changes',
+	Traffic = 'traffic',
+	TrailLength = 'trail-length',
+	TrailLengthMedium = 'trail-length-medium',
+	TrailLengthShort = 'trail-length-short',
+	Train = 'train',
+	Tram = 'tram',
+	Transcribe = 'transcribe',
+	TransferWithinAStation = 'transfer-within-a-station',
+	Transform = 'transform',
+	Transgender = 'transgender',
+	TransitEnterexit = 'transit-enterexit',
+	TransitEnterExit = TransitEnterexit,
+	Translate = 'translate',
+	Transportation = 'transportation',
+	TravelExplore = 'travel-explore',
+	TrendingDown = 'trending-down',
+	TrendingFlat = 'trending-flat',
+	TrendingUp = 'trending-up',
+	TripOrigin = 'trip-origin',
+	Trolley = 'trolley',
+	Tsunami = 'tsunami',
+	Tty = 'tty',
+	Tune = 'tune',
+	TurnLeft = 'turn-left',
+	TurnRight = 'turn-right',
+	TurnSharpLeft = 'turn-sharp-left',
+	TurnSharpRight = 'turn-sharp-right',
+	TurnSlightLeft = 'turn-slight-left',
+	TurnSlightRight = 'turn-slight-right',
+	Tv = 'tv',
+	TvGen = 'tv-gen',
+	TvOff = 'tv-off',
+	TvWithAssistant = 'tv-with-assistant',
+	TwoWheeler = 'two-wheeler',
+	TypeSpecimen = 'type-specimen',
+	UTurnLeft = 'u-turn-left',
+	UTurnRight = 'u-turn-right',
+	UlnaRadius = 'ulna-radius',
+	UlnaRadiusAlt = 'ulna-radius-alt',
+	Umbrella = 'umbrella',
+	Unarchive = 'unarchive',
+	Undo = 'undo',
+	UnfoldLess = 'unfold-less',
+	UnfoldLessDouble = 'unfold-less-double',
+	UnfoldMore = 'unfold-more',
+	UnfoldMoreDouble = 'unfold-more-double',
+	UnknownDocument = 'unknown-document',
+	UnknownMed = 'unknown-med',
+	Unpublished = 'unpublished',
+	Unsubscribe = 'unsubscribe',
+	Upcoming = 'upcoming',
+	Update = 'update',
+	UpdateDisabled = 'update-disabled',
+	Upgrade = 'upgrade',
+	Upload = 'upload',
+	UploadFile = 'upload-file',
+	Urology = 'urology',
+	Usb = 'usb',
+	UsbOff = 'usb-off',
+	Vaccines = 'vaccines',
+	Valve = 'valve',
+	VapeFree = 'vape-free',
+	VapingRooms = 'vaping-rooms',
+	Variables = 'variables',
+	Ventilator = 'ventilator',
+	Verified = 'verified',
+	VerifiedUser = 'verified-user',
+	VerticalAlignBottom = 'vertical-align-bottom',
+	VerticalAlignCenter = 'vertical-align-center',
+	VerticalAlignTop = 'vertical-align-top',
+	VerticalDistribute = 'vertical-distribute',
+	VerticalShades = 'vertical-shades',
+	VerticalShadesClosed = 'vertical-shades-closed',
+	VerticalSplit = 'vertical-split',
+	Vibration = 'vibration',
+	VideoCall = 'video-call',
+	VideoCameraBack = 'video-camera-back',
+	VideoCameraFront = 'video-camera-front',
+	VideoCameraFrontOff = 'video-camera-front-off',
+	VideoChat = 'video-chat',
+	VideoFile = 'video-file',
+	VideoLabel = 'video-label',
+	VideoLibrary = 'video-library',
+	VideoSettings = 'video-settings',
+	VideoStable = 'video-stable',
+	Videocam = 'videocam',
+	VideocamOff = 'videocam-off',
+	VideogameAsset = 'videogame-asset',
+	VideogameAssetOff = 'videogame-asset-off',
+	ViewAgenda = 'view-agenda',
+	ViewArray = 'view-array',
+	ViewCarousel = 'view-carousel',
+	ViewColumn = 'view-column',
+	ViewColumn_2 = 'view-column-2',
+	ViewComfy = 'view-comfy',
+	ViewComfyAlt = 'view-comfy-alt',
+	ViewCompact = 'view-compact',
+	ViewCompactAlt = 'view-compact-alt',
+	ViewCozy = 'view-cozy',
+	ViewDay = 'view-day',
+	ViewHeadline = 'view-headline',
+	ViewInAr = 'view-in-ar',
+	ViewInArNew = 'view-in-ar-new',
+	ViewKanban = 'view-kanban',
+	ViewList = 'view-list',
+	ViewModule = 'view-module',
+	ViewQuilt = 'view-quilt',
+	ViewSidebar = 'view-sidebar',
+	ViewStream = 'view-stream',
+	ViewTimeline = 'view-timeline',
+	ViewWeek = 'view-week',
+	Vignette = 'vignette',
+	Villa = 'villa',
+	Visibility = 'visibility',
+	VisibilityOff = 'visibility-off',
+	VitalSigns = 'vital-signs',
+	Vitals = 'vitals',
+	VoiceChat = 'voice-chat',
+	VoiceOverOff = 'voice-over-off',
+	Voicemail = 'voicemail',
+	Volcano = 'volcano',
+	VolumeDown = 'volume-down',
+	VolumeDownAlt = 'volume-down-alt',
+	VolumeMute = 'volume-mute',
+	VolumeOff = 'volume-off',
+	VolumeUp = 'volume-up',
+	VolunteerActivism = 'volunteer-activism',
+	VotingChip = 'voting-chip',
+	VpnKey = 'vpn-key',
+	VpnKeyOff = 'vpn-key-off',
+	VpnLock = 'vpn-lock',
+	Vrpano = 'vrpano',
+	WallLamp = 'wall-lamp',
+	Wallet = 'wallet',
+	Wallpaper = 'wallpaper',
+	Ward = 'ward',
+	Warehouse = 'warehouse',
+	Warning = 'warning',
+	Wash = 'wash',
+	Watch = 'watch',
+	WatchOff = 'watch-off',
+	Water = 'water',
+	WaterDamage = 'water-damage',
+	WaterDo = 'water-do',
+	WaterDrop = 'water-drop',
+	WaterEc = 'water-ec',
+	WaterHeater = 'water-heater',
+	WaterLux = 'water-lux',
+	WaterOrp = 'water-orp',
+	WaterPh = 'water-ph',
+	WaterVoc = 'water-voc',
+	WaterfallChart = 'waterfall-chart',
+	Waves = 'waves',
+	WavingHand = 'waving-hand',
+	WbAuto = 'wb-auto',
+	WbIncandescent = 'wb-incandescent',
+	WbIridescent = 'wb-iridescent',
+	WbShade = 'wb-shade',
+	WbSunny = 'wb-sunny',
+	WbTwilight = 'wb-twilight',
+	Wc = 'wc',
+	WeatherSnowy = 'weather-snowy',
+	Web = 'web',
+	WebAsset = 'web-asset',
+	WebAssetOff = 'web-asset-off',
+	WebStories = 'web-stories',
+	Webhook = 'webhook',
+	Weekend = 'weekend',
+	Weight = 'weight',
+	West = 'west',
+	Whatshot = 'whatshot',
+	WheelchairPickup = 'wheelchair-pickup',
+	WhereToVote = 'where-to-vote',
+	Widgets = 'widgets',
+	WidthFull = 'width-full',
+	WidthNormal = 'width-normal',
+	WidthWide = 'width-wide',
+	Wifi = 'wifi',
+	Wifi_1Bar = 'wifi-1-bar',
+	Wifi_2Bar = 'wifi-2-bar',
+	WifiCalling = 'wifi-calling',
+	WifiCalling_1 = 'wifi-calling-1',
+	WifiCalling_2 = 'wifi-calling-2',
+	WifiCalling_3 = 'wifi-calling-3',
+	WifiChannel = 'wifi-channel',
+	WifiFind = 'wifi-find',
+	WifiLock = 'wifi-lock',
+	WifiOff = 'wifi-off',
+	WifiPassword = 'wifi-password',
+	WifiProtectedSetup = 'wifi-protected-setup',
+	WifiTethering = 'wifi-tethering',
+	WifiTetheringError = 'wifi-tethering-error',
+	WifiTetheringOff = 'wifi-tethering-off',
+	WindPower = 'wind-power',
+	Window = 'window',
+	WindowClosed = 'window-closed',
+	WindowOpen = 'window-open',
+	WindowSensor = 'window-sensor',
+	WineBar = 'wine-bar',
+	Woman = 'woman',
+	Woman_2 = 'woman-2',
+	Work = 'work',
+	WorkHistory = 'work-history',
+	WorkspacePremium = 'workspace-premium',
+	Workspaces = 'workspaces',
+	WoundsInjuries = 'wounds-injuries',
+	WrapText = 'wrap-text',
+	WrongLocation = 'wrong-location',
+	Wysiwyg = 'wysiwyg',
+	Yard = 'yard',
+	YoutubeActivity = 'youtube-activity',
+	YoutubeSearchedFor = 'youtube-searched-for',
+	ZonePersonAlert = 'zone-person-alert',
+	ZonePersonIdle = 'zone-person-idle',
+	ZonePersonUrgent = 'zone-person-urgent',
+	ZoomIn = 'zoom-in',
+	ZoomInMap = 'zoom-in-map',
+	ZoomOut = 'zoom-out',
+	ZoomOutMap = 'zoom-out-map',
+}
+export type IconSymbolType = `${IconSymbol}` | IconSymbol;
